% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as.data.frame.epiparameter}
\alias{as.data.frame.epiparameter}
\title{\code{\link[=as.data.frame]{as.data.frame()}} method for \verb{<epiparameter>} class}
\usage{
\method{as.data.frame}{epiparameter}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
A \verb{<data.frame>} with a single row.
}
\description{
\code{\link[=as.data.frame]{as.data.frame()}} method for \verb{<epiparameter>} class
}
\details{
The \verb{<data.frame>} returned will contain some atomic columns (i.e. one
object per row), and other columns that are lists (i.e. multiple objects per
row). The list columns can contain lists or S3 objects (e.g. \verb{<bibentry>}
object in the \code{citation} column).
}
\examples{
ep <- epiparameter_db(single_epiparameter = TRUE)
as.data.frame(ep)
}
