% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_db.R
\name{as.data.frame.multi_epidist}
\alias{as.data.frame.multi_epidist}
\title{\link{as.data.frame} method for \verb{<multi_epidist>} class}
\usage{
\method{as.data.frame}{multi_epidist}(x, ...)
}
\arguments{
\item{x}{A \verb{<multi_epidist>} object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A \verb{<data.frame>} with as many rows as length of input list.
}
\description{
\link{as.data.frame} method for \verb{<multi_epidist>} class
}
\details{
The \verb{<data.frame>} returned will contain some atomic columns (i.e. one
object per row), and other columns that are lists (i.e. multiple objects per
row). The list columns can contain lists or S3 objects (e.g. \verb{<bibentry>}
object in the \code{citation} column).
}
