% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{as.function.epidist}
\alias{as.function.epidist}
\title{\code{as.function()} method for \verb{<epidist>} class}
\usage{
\method{as.function}{epidist}(x, func_type = c("density", "cdf", "generate", "quantile"), ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{func_type}{A single \code{character} string specifying which distribution to
convert \verb{<epidist>} object into. Default is \code{"density"}. Other options are
\code{"cdf"}, \code{"generate"}, or \code{"quantile"}.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
A \link{function} object.
}
\description{
Converts an \verb{<epidist>} object to a distribution function
(see \link{epidist_distribution_functions}), either probability density/mass
function, (\code{density}), cumulative distribution function (\code{cdf}), random
number generator (\code{generate}), or quantile (\code{quantile}).
}
\details{
The function returned takes a single required argument \code{x}.
}
