% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist.data.frame}
\alias{as_epidist.data.frame}
\title{Convert \verb{<data.frame>} to an \verb{<epidist>} object}
\usage{
\method{as_epidist}{data.frame}(x, ...)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
An \verb{<epidist>} object or list of \verb{<epidist>} objects.
}
\description{
Convert the tabular information in \verb{<data.frame>} to an \verb{<epidist>}. If the
information in the \verb{<data.frame>} cannot be converted into an \verb{<epidist>} the
function will error.
}
