% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist.dist_spec}
\alias{as_epidist.dist_spec}
\title{Convert a \verb{<dist_spec>} object to an \verb{<epidist>} object}
\usage{
\method{as_epidist}{dist_spec}(x, disease, epi_dist, ...)
}
\arguments{
\item{x}{A \verb{<dist_spec>} object.}

\item{disease}{A \code{character} string with name of the infectious disease.}

\item{epi_dist}{A \code{character} string with the name of the
epidemiological distribution type.}

\item{...}{\link{dots} Extra arguments to be passed to \code{\link[=epidist]{epidist()}}.}
}
\value{
An \verb{<epidist>} object.
}
\description{
Extract distribution parameters and information stored in a \verb{<dist_spec>}
object and store and return in an \verb{<epidist>} object.
}
