% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist.list}
\alias{as_epidist.list}
\title{Convert an \R list to an \verb{<epidist>} object}
\usage{
\method{as_epidist}{list}(x, ...)
}
\arguments{
\item{x}{A list.}

\item{...}{\link{dots} Not used, will warn if extra arguments are provided.}
}
\value{
An \verb{<epidist>} object.
}
\description{
Creates an \verb{<epidist>} object from a list containing valid elements for an
\verb{<epidist>} object. Validation is checked with
\code{\link[=validate_epidist]{validate_epidist()}}.
}
