% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist}
\alias{as_epidist}
\title{Convert an \R object to an \verb{<epidist>} object}
\usage{
as_epidist(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An \verb{<epidist>} object.
}
\description{
Creates an \verb{<epidist>} object from another object with distribution class or
list.
}
