% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{as_epiparam}
\alias{as_epiparam}
\title{Convert an \verb{<epidist>} object to an \verb{<epiparam>} object}
\usage{
as_epiparam(x)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}
}
\value{
An \verb{<epiparam>} object.
}
\description{
Convert an \verb{<epidist>} object to an \verb{<epiparam>} object
}
\examples{
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "lnorm",
  prob_distribution_params = c(meanlog = 1, sdlog = 1)
)
as_epiparam(edist)
}
