% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epiparameter.data.frame}
\alias{as_epiparameter.data.frame}
\title{Convert \verb{<data.frame>} to an \verb{<epiparameter>} object}
\usage{
\method{as_epiparameter}{data.frame}(x, ...)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
An \verb{<epiparameter>} object or list of \verb{<epiparameter>} objects.
}
\description{
Convert the tabular information in \verb{<data.frame>} to an \verb{<epiparameter>}.
If the information in the \verb{<data.frame>} cannot be converted into an
\verb{<epiparameter>} the function will error.
}
