% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{assert_epiparameter}
\alias{assert_epiparameter}
\title{Assert an object is a valid \verb{<epiparameter>} object}
\usage{
assert_epiparameter(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
Invisibly returns an \verb{<epiparameter>}. Called for side-effects
(errors when invalid \verb{<epiparameter>} object is provided).
}
\description{
Assert an object is a valid \verb{<epiparameter>} object
}
