% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{c.epiparameter}
\alias{c.epiparameter}
\title{\code{\link[=c]{c()}} method for \verb{<epiparameter>} class}
\usage{
\method{c}{epiparameter}(...)
}
\arguments{
\item{...}{\link{dots} Objects to be concatenated.}
}
\value{
An \verb{<epiparameter>} or list of \verb{<epiparameter>} objects.
}
\description{
\code{\link[=c]{c()}} method for \verb{<epiparameter>} class
}
\examples{
db <- epiparameter_db()

# combine two <epiparameter> objects into a list
c(db[[1]], db[[2]])

# combine a list of <epiparameter> objects and a single <epiparameter> object
c(db, db[[1]])
}
