% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculates the parametric bootstrap confidence intervals for a data set
simulated under a known probability distribution and parameters}
\usage{
calc_ci(
  data,
  distribution = c("lnorm", "gamma", "weibull"),
  params,
  summary_stat = c("mean", "median", "sd"),
  n_boot = 1000,
  ci = 0.95,
  out_bootci = FALSE
)
}
\arguments{
\item{data}{A numeric vector}

\item{distribution}{A character string, either "lnorm", "gamma", or "weibull"}

\item{params}{The parameters of the distribution}

\item{summary_stat}{A character string, either "mean", "median", "sd"}

\item{n_boot}{A numeric specifying the number of bootstrap iterations}

\item{ci}{A numeric specifying the confidence interval to calculate}

\item{out_bootci}{A boolean determining whether the \code{bootci} class is
returned from the function or just a named vector with the confidence
intervals}
}
\value{
Numeric vector by default, or \code{bootci} object when \code{out_bootci} is
\code{TRUE}
}
\description{
Calculates the parametric bootstrap confidence intervals for a data set
simulated under a known probability distribution and parameters
}
\examples{
dat <- rgamma(n = 100, shape = 1.5, scale = 1)
calc_ci(
  data = dat, 
  distribution = "gamma", 
  params = c(1.5, 1), 
  summary_stat = "mean", 
  n_boot = 1000, 
  ci = 0.95
)
}
