% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_params.R
\name{calc_dist_params}
\alias{calc_dist_params}
\title{Calculates the parameters of a probability distribution from a list of
available summary statistics.}
\usage{
calc_dist_params(prob_dist, summary_stats, sample_size = NULL)
}
\arguments{
\item{prob_dist}{A character string with the name of the probability
distribution}

\item{summary_stats}{A list of summary statistics. The structure of the list
can be easily made with the helper function \code{create_epidist_summary_stats()}.}

\item{sample_size}{The sample size of the data. Only needed when falling back
on using the median-range extraction calculation.}
}
\value{
A named vector with parameters
}
\description{
Calculates the parameters of a probability distribution from a list of
available summary statistics.
}
