% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{$<-.epiparam}
\alias{$<-.epiparam}
\title{Set accessor for \code{epiparam} class}
\usage{
\method{$}{epiparam}(epiparam, ...) <- value
}
\arguments{
\item{epiparam}{An \code{epiparam} object}
}
\value{
An \code{epiparam} object or a \code{data.frame}
}
\description{
Set accessor for \code{epiparam} class
}
