% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_epiparam.R
\name{cbind_epiparam}
\alias{cbind_epiparam}
\title{Bind columns to an \verb{<epiparam>} object}
\usage{
cbind_epiparam(epiparam, ...)
}
\arguments{
\item{epiparam}{An \verb{<epiparam>} object.}

\item{...}{Variables to be added as columns to the \verb{<epiparam>} object.}
}
\value{
An \verb{<epiparam>} object
}
\description{
Adds columns to an existing \verb{<epiparam>} object.

This function avoids the unclassing feature of \code{\link[=cbind]{cbind()}}. The work around
by using \verb{[<-} instead of \code{cbind()} was adapted from
\code{\link[tibble:add_column]{tibble::add_column()}}.
}
\details{
Note that if the variables passed to \code{\link[=cbind_epiparam]{cbind_epiparam()}} are not of
equal length to the number of rows in the \verb{<epiparam>} object then R's
recycling rules apply, see
\url{https://r4ds.had.co.nz/vectors.html#scalars-and-recycling-rules}

Adding columns should not invalidate the \verb{<epiparam>} object, but if it does
the function will throw an error and not return an \verb{<epiparam>} object.
}
\examples{
eparam <- epiparam()
cbind_epiparam(eparam, extra_col = NA)
}
