% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_epidist.R
\name{check_epidist_params}
\alias{check_epidist_params}
\title{Checks the probability distribution name and parameters to be used in an
\code{epidist} object}
\usage{
check_epidist_params(prob_dist, prob_dist_params)
}
\arguments{
\item{prob_dist}{A character string specifying the probability
distribution}

\item{prob_dist_params}{A named vector of probability distribution
parameters}
}
\value{
Nothing, errors if checks fail
}
\description{
Checks the probability distribution name and parameters to be used in an
\code{epidist} object
}
\examples{
epiparameter:::check_epidist_params(
  prob_dist = "gamma",
  prob_dist_params = c(shape = 1, scale = 1)
)
}
\keyword{internal}
