% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_epidist.R
\name{check_epidist_uncertainty}
\alias{check_epidist_uncertainty}
\title{Checks the parameter uncertainty list in an \code{epidist} object}
\usage{
check_epidist_uncertainty(prob_dist_params, uncertainty)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters}

\item{uncertainty}{A list of named vectors with the uncertainty around
the probability distribution parameters. If uncertainty around the parameter
estimates is unknown use \code{create_epidist_uncertainty()} (which is the
argument default) to create a list wiht the correct names with missing
values.}
}
\value{
Nothing, errors if checks fail
}
\description{
Checks the parameter uncertainty list in an \code{epidist} object
}
\examples{
epiparameter:::check_epidist_uncertainty(
  prob_dist_params = c(shape = 1, scale = 1),
  uncertainty = list(
    shape = create_epidist_uncertainty(
      ci = c(0.5, 1),
      ci_interval = 95,
      ci_type = "confidence interval"
    ),
    scale = create_epidist_uncertainty(
      ci = c(0.5, 1),
      ci_interval = 95,
      ci_type = "confidence interval"
    )
  )
)
}
\keyword{internal}
