% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epi_dist}
\alias{clean_epi_dist}
\title{Standardise the names of epidemiological distributions}
\usage{
clean_epi_dist(x)
}
\arguments{
\item{x}{A \code{character} string with the name of the distribution.}
}
\value{
A \code{character} vector of equal length to the input.
}
\description{
Standardise the names of epidemiological distributions
}
\examples{
clean_epi_dist("Incubation_period")
}
