% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_name}
\alias{clean_epidist_name}
\title{Standardise the names of epidemiological distributions}
\usage{
clean_epidist_name(epi_dist)
}
\arguments{
\item{epi_dist}{A character string with the name of the distribution.}
}
\value{
Character string
}
\description{
Standardise the names of epidemiological distributions
}
\examples{
clean_epidist_name("Incubation_period")
}
