% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_params.default}
\alias{clean_epidist_params.default}
\title{Default method if class of parameters is not recognised}
\usage{
\method{clean_epidist_params}{default}(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters.}
}
\value{
Named \code{numeric} vector of parameters.
}
\description{
Default method if class of parameters is not recognised
}
\keyword{internal}
