% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_params.negative_binomial}
\alias{clean_epidist_params.negative_binomial}
\title{Standardises parameters for a negative binomial distribution}
\usage{
\method{clean_epidist_params}{negative_binomial}(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters}
}
\value{
Named vector of parameters
}
\description{
Standardises parameters for a negative binomial distribution
}
\keyword{internal}
