% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_params.pois}
\alias{clean_epidist_params.pois}
\title{Standardise parameters for a poisson distribution}
\usage{
\method{clean_epidist_params}{pois}(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters.}
}
\value{
Named vector of parameters
}
\description{
Standardise parameters for a poisson distribution
}
\keyword{internal}
