% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_params}
\alias{clean_epidist_params}
\title{Dispatches to a parameter cleaning function depending on distribution}
\usage{
clean_epidist_params(prob_dist, prob_dist_params)
}
\arguments{
\item{prob_dist}{A character string specifying the probability
distribution. This should match the R naming convention of probability
distributions (e.g. lognormal is lnorm, negative binomial is nbinom, and
geometric is geom).}

\item{prob_dist_params}{A named vector of probability distribution
parameters.}
}
\value{
Named \code{numeric} vector of parameters.
}
\description{
Dispatches to a parameter cleaning function depending on distribution
}
\keyword{internal}
