% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_epidist_params.weibull}
\alias{clean_epidist_params.weibull}
\title{Standardise parameters for a Weibull distribution}
\usage{
\method{clean_epidist_params}{weibull}(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters}
}
\value{
Named vector of parameters
}
\description{
Standardise parameters for a Weibull distribution
}
\keyword{internal}
