% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_gamma_params}
\alias{convert_gamma_params}
\title{Converts the parameters of the gamma distribution to summary statistics}
\usage{
convert_gamma_params(shape, scale)
}
\arguments{
\item{shape}{The shape parameter of the gamma distribution}

\item{scale}{The scale parameter of the gamma distribution}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and excess kurtosis (\code{ex_kurtosis}).
}
\description{
Converts the shape and scale parameters of the gamma
distribution to a number of summary statistics which can be calculated
analytically given the gamma parameters. One exception is the median which
is calculated using \code{\link[=qgamma]{qgamma()}} as no analytical form is available.
}
\examples{
convert_gamma_params(shape = 1, scale = 1)
}
