% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_geom_params}
\alias{convert_geom_params}
\title{Converts the parameters of the geometric distribution to summary statistics}
\usage{
convert_geom_params(prob)
}
\arguments{
\item{prob}{The probability parameter (p) of the geometric distribution.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and excess kurtosis (\code{ex_kurtosis}).
}
\description{
Converts the probability (\code{prob}) of the geometric distribution
to a number of summary statistics which can be calculated analytically given
the geometric parameter.
}
\details{
This conversion function assumes that distribution represents the
number of failures before the first success (supported for zero). This is
the same form as used by base R and \code{distributional::dist_geometric()}.
}
\examples{
convert_geom_params(prob = 1)
}
