% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_lnorm_params}
\alias{convert_lnorm_params}
\title{Convert the parameters of the lognormal distribution to summary statistics}
\usage{
convert_lnorm_params(meanlog, sdlog)
}
\arguments{
\item{meanlog}{The meanlog parameter (mean of natural logarithm) of the
lognormal distribution. Also commonly called and denoted by the mu.}

\item{sdlog}{The sdlog parameter (standard deviation of the natural
logarithm) of the distribution. Also commonly called and denoted by sigma.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and excess kurtosis (\code{ex_kurtosis}).
}
\description{
Converts the meanlog and sdlog parameters of the lognormal
distribution to a number of summary statistics which can be calculated
analytically given the lognormal parameters.
}
\examples{
convert_lnorm_params(meanlog = 1, sdlog = 1)
}
