% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_nbinom_params}
\alias{convert_nbinom_params}
\title{Converts the parameters of the negative binomial distribution to summary
statistics}
\usage{
convert_nbinom_params(prob, dispersion)
}
\arguments{
\item{prob}{The probability parameter (p) of the negative binomial}

\item{dispersion}{The dispersion parameter (k) of the negative binomial. This
parameter is also commonly represented as \emph{r}.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and kurtosis.
}
\description{
Converts the probability (\code{prob}) and dispersion parameters of
the negative binomial distribution to a number of summary statistics which
can be calculated analytically given the negative binomial parameters.
One exception is the median which is calculated using \code{\link[=qnbinom]{qnbinom()}} as no
analytical form is available.
}
\examples{
convert_nbinom_params(prob = 1, dispersion = 1)
}
