% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_nbinom_summary_stats}
\alias{convert_nbinom_summary_stats}
\title{Converts the summary statistics to parameters of the negative binomial
distribution}
\usage{
convert_nbinom_summary_stats(...)
}
\arguments{
\item{...}{Summary statistics to be used for calculating the negative
binomial distribution parameters. All arguments must be correctly named,
see details for possible input.}
}
\value{
A list of two elements, the probability and size parameters
}
\description{
Converts the summary statistics of the negative binomial
distribution the parameters (\code{prob}) and (\code{size}) of the negative binomial
distribution.
}
\details{
The arguments input are captured and handled by name. Therefore,
the names of the arguments must be correct (case-sensitive). Possible input
summary statistics are: \code{mean}, \code{median}, \code{mode}, \code{var}, \code{sd}, \code{cv},
\code{skewness}, \code{kurtosis}.

Not every combination of input statistics is a viable conversion, and may
error. Common conversions that are implemented are mean and any measure of
spread (\code{var}, \code{sd} and \code{cv}), or \code{median} and \code{sd}.
}
\examples{
convert_nbinom_summary_stats(mean = 3)
}
