% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_params.R
\name{convert_params.gamma}
\alias{convert_params.gamma}
\alias{convert_params.lognormal}
\alias{convert_params.weibull}
\title{Converts summary statistics to parameters of a probability distribution}
\usage{
\method{convert_params}{gamma}(summary_stats)

\method{convert_params}{lognormal}(summary_stats)

\method{convert_params}{weibull}(summary_stats)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviation}
}
\value{
Named vector
}
\description{
The summary_stats object needs to have a class attribute with
the name of the probability distribution in order for the correct conversion
function to be dispatched.
}
\keyword{internal}
