% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_params_weibull}
\alias{convert_params_weibull}
\title{Converts the parameters of the Weibull distribution to summary statistics}
\usage{
convert_params_weibull(...)
}
\arguments{
\item{...}{\code{Numeric} named parameter(s) used to convert to summary
statistics. An example is the meanlog and sdlog parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and excess kurtosis (\code{ex_kurtosis}).
}
\description{
Converts the shape and scale parameters of the Weibull
distribution to a number of summary statistics which can be calculated
analytically given the Weibull parameters. Note the conversion uses the
\code{\link[=gamma]{gamma()}} function.
}
\keyword{internal}
