% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_summary_stats_gamma}
\alias{convert_summary_stats_gamma}
\title{Convert summary statistics to parameters of the gamma distribution}
\usage{
convert_summary_stats_gamma(...)
}
\arguments{
\item{...}{\link{dots} \code{Numeric} named summary statistics used to convert to
parameter(s). An example is the \code{mean} and \code{sd} parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of two elements, the shape and scale
}
\description{
Convert the summary statistics input into the shape and scale
parameters of the gamma distribution.
}
\keyword{internal}
