% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_summary_stats_geom}
\alias{convert_summary_stats_geom}
\title{Convert summary statistics to parameters of the geometric distribution}
\usage{
convert_summary_stats_geom(...)
}
\arguments{
\item{...}{\code{Numeric} named summary statistics used to convert to
parameter(s). An example is the meanlog and sdlog parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of one element, the probability parameter.
}
\description{
Convert summary statistics of the geometric
distribution the parameter (\code{prob}) of the geometric distribution.
}
\details{
This conversion function assumes that distribution represents the
number of failures before the first success (supported for zero). This is
the same form as used by base R and \code{\link[distributional:dist_geometric]{distributional::dist_geometric()}}.
}
\keyword{internal}
