% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_summary_stats_lnorm}
\alias{convert_summary_stats_lnorm}
\title{Converts the summary statistics to parameters of the lognormal distribution}
\usage{
convert_summary_stats_lnorm(...)
}
\arguments{
\item{...}{\code{Numeric} named summary statistics used to convert to
parameter(s). An example is the meanlog and sdlog parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of two elements, the meanlog and sdlog
}
\description{
Converts the summary statistics input into the meanlog and sdlog
parameters of the lognormal distribution.
}
\keyword{internal}
