% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_summary_stats_nbinom}
\alias{convert_summary_stats_nbinom}
\title{Converts the summary statistics to parameters of the negative binomial
distribution}
\usage{
convert_summary_stats_nbinom(...)
}
\arguments{
\item{...}{\code{Numeric} named summary statistics used to convert to
parameter(s). An example is the meanlog and sdlog parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of two elements, the probability and dispersion parameters
}
\description{
Converts the summary statistics of the negative binomial
distribution the parameters (\code{prob}) and (\code{dispersion}) of the negative
binomial distribution.
}
\keyword{internal}
