% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_metadata}
\alias{create_epidist_metadata}
\title{A helper function when creating an epidist object to create a metadata list
with sensible defaults, type checking and arguments to help remember metadata
list structure (element names)}
\usage{
create_epidist_metadata(
  sample_size = NULL,
  region = NULL,
  vector_borne = FALSE,
  vector = NULL,
  inference_method = NULL
)
}
\arguments{
\item{sample_size}{stub}

\item{region}{stub}

\item{vector_borne}{stub}

\item{vector}{stub}

\item{inference_method}{stub}
}
\value{
named list
}
\description{
A helper function when creating an epidist object to create a metadata list
with sensible defaults, type checking and arguments to help remember metadata
list structure (element names)
}
\examples{
a = 1
}
