% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_method_assessment}
\alias{create_epidist_method_assessment}
\title{A helper function when creating an epidist object to create a method
assessment list with sensible defaults, type checking and arguments to help
remember which method assessments can be accepted in the list}
\usage{
create_epidist_method_assessment(
  censorred = NULL,
  right_truncated = NULL,
  phase_bias_adjusted = NULL
)
}
\arguments{
\item{censorred}{A boolean logical whether the study used single or double
interval censorring in the methods to infer the delay distribution}

\item{right_truncated}{A boolean logical whether the study used right-
truncation in the methods to infer the delay distribution}

\item{phase_bias_adjusted}{A boolean logical whether the study adjusted for
phase bias in the methods to infer the delay distribution}
}
\value{
A list of three elements
}
\description{
A helper function when creating an epidist object to create a method
assessment list with sensible defaults, type checking and arguments to help
remember which method assessments can be accepted in the list
}
\examples{
create_epidist_method_assessment(
  censorred = FALSE,
  right_truncated = FALSE,
  phase_bias_adjusted = FALSE
)
}
