% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_region}
\alias{create_epidist_region}
\title{Specify the geography of the data entry}
\usage{
create_epidist_region(
  continent = NA_character_,
  country = NA_character_,
  region = NA_character_,
  city = NA_character_
)
}
\arguments{
\item{continent}{A character string specifying the continent.}

\item{country}{A character string specifying the country.}

\item{region}{A character string specifying the region.}

\item{city}{A character string specifying the city.}
}
\value{
A named list
}
\description{
The geography of the data set can be a single geographical
region at either continent, country, region or city level. By specifying
the level of the geography the other fields may be deduced.
}
\examples{
create_epidist_region(country = "UK")
}
