% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_summary_stats}
\alias{create_epidist_summary_stats}
\title{A helper function when creating an epidist object to create a summary
statistics list with sensible defaults, type checking and arguments to help
remember which summary statistics can be accepted in the list}
\usage{
create_epidist_summary_stats(
  mean = NULL,
  mean_ci = NULL,
  mean_ci_interval = NULL,
  sd = NULL,
  sd_ci = NULL,
  sd_ci_interval = NULL,
  median = NULL,
  median_ci = NULL,
  median_ci_interval = NULL,
  dispersion = NULL,
  dispersion_ci = NULL,
  dispersion_ci_interval = NULL,
  lower_range = NULL,
  upper_range = NULL,
  q_025 = NULL,
  q_05 = NULL,
  q_25 = NULL,
  q_75 = NULL,
  q_875 = NULL,
  q_95 = NULL,
  q_975 = NULL
)
}
\arguments{
\item{mean}{A numeric of the mean (expectation) of the probability
distribution}

\item{mean_ci}{A numeric vector of length two of the confidence interval
around the mean}

\item{mean_ci_interval}{A numeric specifying the confidence interval width,
e.g. 95 would be the 95\% CI}

\item{sd}{A numeric of the standard deviation of the probability distribution}

\item{sd_ci}{A numeric vector of length 2 of the confidence interval around
the standard deviation}

\item{sd_ci_interval}{A numeric specifying the confidence interval width,
e.g. 95 would be 95\% confidence interval}

\item{median}{A numeric of the median of the probability distribution}

\item{median_ci}{A numeric vector of length two of the confidence interval
around the median}

\item{median_ci_interval}{A numeric specifying the confidence interval width
of the median}

\item{dispersion}{A numeric of the dispersion parameter of a distribution}

\item{dispersion_ci}{A numeric vector of length two of the confidence
interval around the dispersion}

\item{dispersion_ci_interval}{A numeric specifying the confidence interval
width of the dispersion parameter}

\item{lower_range}{The lower range of the data, used to infer the parameters
of the distribution when not provided}

\item{upper_range}{The upper range of the data, used to infer the parameters
of the distribution when not provided}

\item{q_025}{The 2.5th quantile of the probability distribution}

\item{q_05}{The 5th quantile of the probability distribution}

\item{q_25}{The 25th quantile of the probability distribution}

\item{q_75}{The 75th quantile of the probability distribution}

\item{q_875}{The 87.5th quantile of the probability distribution}

\item{q_95}{The 95th quantile of the probability distribution}

\item{q_975}{The 97.5th quantile of the probability distribution}
}
\value{
A nested list of summary statistics. The highest level are
\itemize{
\item \verb{$central_tendency_spread}
\item \verb{$quantiles}
\item \verb{$range}
\item \verb{$dispersion}
}
}
\description{
A helper function when creating an epidist object to create a summary
statistics list with sensible defaults, type checking and arguments to help
remember which summary statistics can be accepted in the list
}
\examples{
create_epidist_summary_stats(mean = 5, sd = 2)
}
