% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{discretise}
\alias{discretise}
\alias{discretise.epidist}
\alias{discretise.default}
\title{Discretises a continuous distribution in an \code{epidist} object}
\usage{
discretise(x, ...)

\method{discretise}{epidist}(x, ...)

\method{discretise}{default}(x, ...)
}
\arguments{
\item{x}{An \code{epidist} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
An \code{epidist} object
}
\description{
Discretises a continuous distribution in an \code{epidist} object
}
\details{
Converts the S3 distribution object in an \code{epidist} from continuous
(using an object from the {distributional} package) to a discretised
distribution (using an object from the {distcrete} package).
}
\examples{
ebola_incubation <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
discretise(ebola_incubation)
}
