% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.citet}
\alias{.citet}
\title{Format short citation from \verb{<bibentry>} object}
\usage{
.citet(x)
}
\arguments{
\item{x}{A \verb{<bibentry>} object, see \code{\link[=bibentry]{bibentry()}}.}
}
\value{
A \code{character} string with the short citation
}
\description{
Output is equivalent to the \verb{\\citet\{\}} function in the \pkg{natbib} LaTeX
package.
}
\keyword{internal}
