% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{.clean_epidist_params_nbinom}
\alias{.clean_epidist_params_nbinom}
\title{Standardise parameters for a negative binomial distribution}
\usage{
.clean_epidist_params_nbinom(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters.}
}
\value{
Named \code{numeric} vector of parameters.
}
\description{
Standardise parameters for a negative binomial distribution
}
\keyword{internal}
