% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter-utils.R
\name{.clean_uncertainty}
\alias{.clean_uncertainty}
\title{Standardise distribution parameter uncertainty}
\usage{
.clean_uncertainty(x, prob_distribution, uncertainty_missing)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{prob_distribution}{An S3 class containing the probability
distribution or a character string if the parameters of the probability
distribution are unknown but the name of the distribution is known, or \code{NA}
if the distribution name and parameters are unknown. Use
\code{\link[=create_prob_distribution]{create_prob_distribution()}} to create \code{prob_distribution}.}

\item{uncertainty_missing}{A boolean \code{logical} as to whether \code{uncertainty}
is missing (see \code{\link[=missing]{missing()}}) from the parent function.}
}
\value{
An uncertainty list for an \verb{<epiparameter>} object.
}
\description{
Standardise distribution parameter uncertainty
}
\keyword{internal}
