% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{.convert_summary_stats_lnorm}
\alias{.convert_summary_stats_lnorm}
\title{Convert summary statistics to parameters of the lognormal distribution}
\usage{
.convert_summary_stats_lnorm(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> \code{Numeric} named summary
statistics used to convert to parameter(s). An example is the \code{mean}
and \code{sd} summary statistics for the lognormal (\code{lnorm}) distribution.}
}
\value{
A list of two elements: meanlog and sdlog
}
\description{
Convert the summary statistics input into the meanlog and sdlog
parameters of the lognormal distribution.
}
\keyword{internal}
