% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.study_key}
\alias{.study_key}
\title{Key identifying the disease, parameter and study of an \verb{<epiparameter>}}
\usage{
.study_key(x, i)
}
\arguments{
\item{x}{A \verb{<multi_epiparameter>} object.}

\item{i}{The index of \code{x} in the \verb{<multi_epiparameter>}, used to keep
entries without an identifiable study from being grouped together.}
}
\value{
A \code{character} string.
}
\description{
Key identifying the disease, parameter and study of an \verb{<epiparameter>}
}
\keyword{internal}
