% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_db.R
\name{epidist_db}
\alias{epidist_db}
\title{Create an \code{epidist} object directly from the epiparameter library (database)}
\usage{
epidist_db(
  disease,
  epi_dist = c("incubation_period", "onset_to_hospitalisation", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  author = NULL
)
}
\arguments{
\item{disease}{A character string specifying the disease}

\item{epi_dist}{A character string specifying the epidemiological
distribution}

\item{author}{The author of the study reporting the distribution}
}
\value{
An \code{epidist} object
}
\description{
This function can extract an \code{epidist} object directly from the
library of epidemiological parameters without having to read in an \code{epiparam}
object and pull out an \code{epidist} object from one of the entries. If a
distribution from a specific study is required, the \code{author} argument can
be specified. For now this must match the author entry in the database
exactly to be recognised.
}
\details{
The current criteria for chosing a distribution when multiple
entries exist in the database for a given disease and epidemiological
distribution is to return the entry with the largest sample size.
}
\examples{
epidist_db(disease = "influenza", epi_dist = "serial_interval")

# comparison between using `epidist_db()` and `epiparam()` with
# `as_epidist()`

# load influenza serial interval from database
edist <- epidist_db(disease = "influenza", epi_dist = "serial_interval")

# load database of serial intervals
eparam <- epiparam(epi_dist = "serial_interval")
# subset database to only influenza entries
eparam <- eparam[clean_disease(eparam$disease) == "influenza", ]
# convert to `epidist`
edist2 <- as_epidist(eparam)
# check the two methods produce the same `epidist` object
identical(edist, edist2)
}
