% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{epidist_df_to_epidist}
\alias{epidist_df_to_epidist}
\title{Convert \verb{<data.frame>} from \code{\link[=as.data.frame.epidist]{as.data.frame.epidist()}} to \verb{<epidist>}}
\usage{
epidist_df_to_epidist(x, ...)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}

\item{...}{\link{dots} Extra arguments to pass to \code{\link[=epidist]{epidist()}}.}
}
\value{
An \verb{<epidist>} object.
}
\description{
Convert \verb{<data.frame>} from \code{\link[=as.data.frame.epidist]{as.data.frame.epidist()}} to \verb{<epidist>}
}
\keyword{internal}
