% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{epiparam}
\alias{epiparam}
\title{Create an \verb{<epiparam>} object}
\usage{
epiparam(
  epi_dist = c("all", "incubation_period", "onset_to_hospitalisation", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution")
)
}
\arguments{
\item{epi_dist}{A \code{character} string with the name of the
epidemiological distribution type.}
}
\value{
An \verb{<epiparam>} object.
}
\description{
The \verb{<epiparam>} class holds information on epidemiological
distribution and their estimated parameters as well as other information and
metadata. This library of epidemiological parameters is compiled from
primary literature sources. An \verb{<epiparam>} object can be used to compare the
availability of distribution for a certain disease or pathogen, or refine
by, for example, region or sample size. Additionally, the \verb{<epiparam>} class
can be subset and converted into \verb{<epidist>} or \verb{<vb_epidist>} objects to
the be used in epidemiological analysis in which a delay distribution or
offspring distribution is required.

The \code{\link[=epiparam]{epiparam()}} function reads the library of epidemiological parameters
from \code{{epiparameter}} into memory and stores it as an \verb{<epiparam>} object.
}
\details{
The \verb{<epiparam>} object has certain protected fields, and thus if
one of these protected fields is removed when subsetting columns an error
will be returned. The subsetting checks are carried out by
\code{\link[=validate_epiparam]{validate_epiparam()}}.

Data can be added to \verb{<epiparam>} objects by using \code{\link[=bind_epiparam]{bind_epiparam()}}, this
can add information from \verb{<epidist>}, \verb{<vb_epdist>}, \verb{<epiparam>}, lists
of \verb{<epdist>} objects, or data frames with the correct columns to an
existing \verb{<epiparam>} object.
}
\examples{
# the object can be made without arguments
eparam <- epiparam()

# specifying incubation periods
incub_eparam <- epiparam("incubation")

# subset by disease
influenza_dists <- eparam[eparam$disease == "influenza", ]
}
