% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{epiparameter_df_to_epiparameter}
\alias{epiparameter_df_to_epiparameter}
\title{Convert \verb{<data.frame>} from \code{\link[=as.data.frame.epiparameter]{as.data.frame.epiparameter()}} to
\verb{<epiparameter>}}
\usage{
epiparameter_df_to_epiparameter(x, ...)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}

\item{...}{\link{dots} Extra arguments to pass to \code{\link[=epiparameter]{epiparameter()}}.}
}
\value{
An \verb{<epiparameter>} object.
}
\description{
Convert \verb{<data.frame>} from \code{\link[=as.data.frame.epiparameter]{as.data.frame.epiparameter()}} to
\verb{<epiparameter>}
}
\keyword{internal}
