% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epireview_core_cols}
\alias{epireview_core_cols}
\title{A vector of \code{character} strings with the core column names of the
epidemiological parameter data exported by the \pkg{epireview} R package.}
\format{
\subsection{\code{epireview_core_cols}}{

A \code{character} vector with 21
elements.

These are the columns of the \pkg{epireview} epidemiological parameter table
(\verb{load_epidata(<pathogen>)$params}) that are required to convert an
\pkg{epireview} entry into an \verb{<epiparameter>} object with
\code{\link[=as_epiparameter]{as_epiparameter()}}. They are used to identify an \pkg{epireview} parameter
table and to check that the columns needed for conversion are present.
}
}
\source{
\url{https://github.com/mrc-ide/epireview}
}
\usage{
epireview_core_cols
}
\description{
A vector of \code{character} strings with the core column names of the
epidemiological parameter data exported by the \pkg{epireview} R package.
}
\keyword{datasets}
