% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{epireview_to_epidist}
\alias{epireview_to_epidist}
\title{Convert \verb{<data.frame>} from \pkg{epireview} to \verb{<epidist>}}
\usage{
epireview_to_epidist(x, ...)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}

\item{...}{\link{dots} Extra arguments to pass to \code{\link[=epidist]{epidist()}}.}

\item{article}{A \verb{<data.frame>} with bibliographic information on the
epidemiological parameters. See Details for more information.}
}
\value{
An \verb{<epidist>} object.
}
\description{
Convert \verb{<data.frame>} from \pkg{epireview} to \verb{<epidist>}
}
\details{
To create the full citation the information from the article table from the
\pkg{epireview} package of the corresponding entry will need to be passed
to function via the \code{...} argument. The argument should be called \code{article},
as it will be matched by name by \code{$}.
}
\keyword{internal}
