% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{Extract parameters}
\usage{
extract_param(type, values, distribution)
}
\arguments{
\item{type}{A \code{character} defining whether summary statistics based around \code{median} and 95\% CI (default) or \code{mean} and variance.}

\item{values}{A \code{vector} with three elements if \code{type=median}: c(median, lower 25\% CI, upper 75\% CI); and with two elements if \code{type=mean}: c(mean, variance)}

\item{distribution}{A \code{character} specifying distribution to use. Default is \code{lognormal}; also takes \code{gamma}.}
}
\description{
Extract parameters from summary data (e.g. in meta analysis)
}
\examples{
extract_param()
}
\keyword{extract}
