% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param_percentile}
\alias{extract_param_percentile}
\title{Optimises the parameters for a specified probability distribution given the
percentiles of a distribution and the values at those percentiles}
\usage{
extract_param_percentile(values, distribution, percentiles)
}
\arguments{
\item{values}{A \code{vector}. If \code{type = percentiles}: \code{c(percentile_1, percentile_2)}; and if \code{type = range}: \code{c(median, min, max)}}

\item{distribution}{A \code{character} specifying distribution to use.
Default is \code{lnorm}; also takes \code{gamma} and \code{weibull}.}

\item{percentiles}{A \code{vector} with two elements specifying the
percentiles defined in \code{values} if using \code{type = "percentiles"}.}
}
\value{
A list with output from stats::optim. See ?optim for more details
}
\description{
Optimises the parameters for a specified probability distribution given the
percentiles of a distribution and the values at those percentiles
}
\examples{
# extract parameters of a lognormal distribution from the 75 percentiles
extract_param_percentile(
  values = c(6, 13),
  distribution = "lnorm",
  percentiles = c(0.125, 0.875)
)
}
\keyword{internal}
