% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{format.epiparam}
\alias{format.epiparam}
\title{Format method for \verb{<epiparam>} class}
\usage{
\method{format}{epiparam}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparam>} object.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
Invisibly returns an \verb{<epiparam>}. Called for printing side-effects.
}
\description{
Format method for \verb{<epiparam>} class
}
\examples{
x <- epiparam()
format(x)
}
