% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{format.epiparameter}
\alias{format.epiparameter}
\title{Format method for \verb{<epiparameter>} class}
\usage{
\method{format}{epiparameter}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
Invisibly returns an \verb{<epiparameter>}. Called for printing
side-effects.
}
\description{
Format method for \verb{<epiparameter>} class
}
\examples{
epiparameter <- epiparameter(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = create_prob_distribution(
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1)
  )
)
format(epiparameter)
}
