% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{Gamma}
\alias{Gamma}
\title{Constructor for \verb{<gamma_dist>} object}
\usage{
Gamma(shape, rate = 1, scale = 1/rate, log = FALSE, ..., mean, sd)
}
\arguments{
\item{shape, scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{rate}{an alternative way to specify the scale.}
}
\value{
A \verb{<gamma_dist>} object.
}
\description{
Constructor for \verb{<gamma_dist>} object
}
\details{
Parameterising the gamma distribution with mean and standard deviation does
not allow specifying parameter uncertainty. If \code{mean} or \code{sd} are specified
as anything other than a single \code{numeric} the function will error.
}
