% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{gamma_meansd2shapescale}
\alias{gamma_meansd2shapescale}
\title{Converts the mean and standard deviation of the gamma distribution to the
shape and scale parameterisation.}
\usage{
gamma_meansd2shapescale(mean, sd)
}
\arguments{
\item{mean}{The mean of the gamma distribution}

\item{sd}{The standard deviation of the gamma distribution}
}
\value{
A named list with shape and scale parameters
}
\description{
Converts the mean and standard deviation of the gamma distribution to the
shape and scale parameterisation.
}
\examples{
gamma_meansd2shapescale(mean = 2.2, sd = 0.9)
}
