% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{gamma_shapescale2meansd}
\alias{gamma_shapescale2meansd}
\title{Converts the shape and scale parameters of the gamma distribution to the
mean and standard deviation.}
\usage{
gamma_shapescale2meansd(shape, scale)
}
\arguments{
\item{shape}{A single numeric of the shape parameter of the gamma
distribution}

\item{scale}{A single numeric of the scale parameter of the gamma
distribution}
}
\value{
A named list with mean and standard deviation
}
\description{
Converts the shape and scale parameters of the gamma distribution to the
mean and standard deviation.
}
\examples{
gamma_shapescale2meansd(shape = 0.5, scale = 0.2)
}
