% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation.epidist}
\alias{get_citation.epidist}
\title{Get citation from an \verb{<epidist>} object}
\usage{
\method{get_citation}{epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
A \verb{<bibentry>} object.
}
\description{
Extract the citation stored in an \verb{<epidist>} object.
}
\examples{
# example with <epidist>
edist <- epidist_db(single_epidist = TRUE)
get_citation(edist)

# example returning bibtex format
edist <- epidist_db(disease = "COVID-19", single_epidist = TRUE)
cit <- get_citation(edist)
format(cit, style = "bibtex")
}
