% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation.multi_epidist}
\alias{get_citation.multi_epidist}
\title{Get citation from a list of \verb{<epidist>} objects}
\usage{
\method{get_citation}{multi_epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
A \verb{<bibentry>} object containing multiple references. The length of
output \verb{<bibentry>} is equal to the length of the list of \verb{<epidist>}
objects supplied.
}
\description{
Extract the citation stored in a list of \verb{<epidist>} objects.
}
\examples{
# example with list of <epidist>
edist <- epidist_db()
get_citation(edist)
}
