% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation.multi_epiparameter}
\alias{get_citation.multi_epiparameter}
\title{Get citation from a list of \verb{<epiparameter>} objects}
\usage{
\method{get_citation}{multi_epiparameter}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
A \verb{<bibentry>} object containing multiple references. The length of
output \verb{<bibentry>} is equal to the length of the list of \verb{<epiparameter>}
objects supplied.
}
\description{
Extract the citation stored in a list of \verb{<epiparameter>} objects.
}
\examples{
# example with list of <epiparameter>
db <- epiparameter_db()
get_citation(db)
}
