% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_parameters.epidist}
\alias{get_parameters.epidist}
\title{Get parameters from an \verb{<epidist>} object}
\usage{
\method{get_parameters}{epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{Extra arguments to be passed to the method.}
}
\value{
A named vector of parameters or \code{NA} when the \verb{<epidist>} object is
unparameterised.
}
\description{
Extract the parameters of the distribution stored in an \verb{<epidist>} object.
}
\details{
The \verb{<epidist>} object can be unparameterised in which it lacks
a probability distribution or parameters of a probability distribution,
in this case the \code{get_parameters.epidist()} method will return \code{NA}.
}
\examples{
edist <- epidist_db(
  disease = "COVID-19",
  epi_dist = "serial interval",
  single_epidist = TRUE
)
get_parameters(edist)
}
