% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from an object}
\usage{
get_parameters(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
A named vector of parameters or \code{NA} when the \verb{<epiparameter>}
object is unparameterised, or a list where each element is a named vectors
or \code{NA}.
}
\description{
Extract the parameters stored in an \R object.
}
