% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentiles.R
\name{get_percentiles}
\alias{get_percentiles}
\title{Converts a vector of named percentiles into correct format and selects two
values for parameter extraction}
\usage{
get_percentiles(percentiles)
}
\arguments{
\item{percentiles}{A named vector of values at percentiles with the names the
percentiles. See Details for the accepted vector name format.}
}
\value{
A named vector of percentiles
}
\description{
Parameters of a probability distribution can be extracted using
the values given at percentiles of the distribution and the percentiles using
\code{\link[=extract_param]{extract_param()}}. \code{get_percentiles()} takes a named vector of percentiles
(names) and values at those percentiles (elements in the vector) and selects
two values as lower and upper percentiles to be used in extraction.

It also formats the vector names so that they can be correctly converted to
numeric using \code{as.numeric()}.
}
\details{
The name format is "q_" followed by the value. Decimal points
are not used an should have a zero infront of the number to represent
decimal places, e.g. q_05 is 0.5. Numbers with decimal places after non-zero
values are given as q_975 which is 97.5.
}
\examples{
\dontrun{
  # 90th interval
  get_percentiles(c(q_05 = 1, q_95 = 10))
  # 95th interval
  get_percentiles(c(q_025 = 1, q_975 = 10))
}
}
\keyword{internal}
