% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{head.epiparam}
\alias{head.epiparam}
\alias{tail.epiparam}
\title{\code{\link[=head]{head()}} and \code{\link[=tail]{tail()}} methods for \verb{<epiparam>} class}
\usage{
\method{head}{epiparam}(x, ...)

\method{tail}{epiparam}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparam>} object.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
Data frame
}
\description{
\code{\link[=head]{head()}} and \code{\link[=tail]{tail()}} methods for \verb{<epiparam>} class
}
\examples{
head(epiparam())
tail(epiparam())
}
