% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{head}
\alias{head}
\title{head method for epiparam class}
\usage{
head(x)
}
\arguments{
\item{x}{epiparam object}
}
\value{
Data frame
}
\description{
head method for epiparam class
}
\examples{
head(epiparam())
}
