% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incubation_f.R
\name{incubation_f}
\alias{incubation_f}
\title{Incubation period}
\usage{
incubation_f(pathogen, study, pmf)
}
\arguments{
\item{pathogen}{A \code{character} specifying pathogen of interest}

\item{study}{A \code{character} specifying dataset to use. Defaults to study with largest sample size.}

\item{pmf}{A \code{logical} specifying whether to output probability mass function (\code{TRUE}) or cumulative mass function (\code{FALSE}). Defaults to \code{TRUE}.}
}
\description{
Incubation period probability distribution (i.e. exposure-to-onset by day)
}
\examples{
incubation_f()
}
\keyword{incubation}
