% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{is_epiparam}
\alias{is_epiparam}
\title{Check object is an \verb{<epiparam>}}
\usage{
is_epiparam(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A boolean \code{logical}, \code{TRUE} if the object is an \verb{<epiparam>} and
\code{FALSE} if not.
}
\description{
Check object is an \verb{<epiparam>}
}
\examples{
eparam <- epiparam()

is_epiparam(eparam)
}
