% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{is_epiparameter_df}
\alias{is_epiparameter_df}
\title{Check if \verb{<data.frame>} input is from \code{\link[=as.data.frame.epiparameter]{as.data.frame.epiparameter()}}}
\usage{
is_epiparameter_df(x)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}
}
\value{
A single \code{logical} boolean.
}
\description{
Check if \verb{<data.frame>} input is from \code{\link[=as.data.frame.epiparameter]{as.data.frame.epiparameter()}}
}
\keyword{internal}
