% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{is_epireview}
\alias{is_epireview}
\title{Check if \verb{<data.frame>} input is from \pkg{epireview}}
\usage{
is_epireview(x)
}
\arguments{
\item{x}{A \verb{<data.frame>}.}
}
\value{
A single \code{logical} boolean.
}
\description{
Check if \verb{<data.frame>} input is from \pkg{epireview}
}
\keyword{internal}
