% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{is_truncated}
\alias{is_truncated}
\title{Check if distribution in \code{epidist} is truncated}
\usage{
is_truncated(x)
}
\arguments{
\item{x}{An \code{epidist} object.}
}
\value{
A boolean logical.
}
\description{
Check if distribution in \code{epidist} is truncated
}
\details{
The \verb{<epidist>} class can hold probability distribution objects
from the {distributional} package or the {distcrete} package, however, only
distribution objects from {distributional} can be truncated. If a
\verb{<epidist>} object has a {distcrete} object \code{is_truncated} will return
\code{FALSE} by default.
}
\examples{
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "lnorm",
  prob_distribution_params = c(meanlog = 1, sdlog = 1)
)
is_truncated(edist)

edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "lnorm",
  prob_distribution_params = c(meanlog = 1, sdlog = 1),
  truncation = 10
)
is_truncated(edist)
}
