% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{is_valid_vector_borne}
\alias{is_valid_vector_borne}
\title{Checks whether the data input into making an \code{epidist} class object is valid
for a vector-borne disease with two delay distributions (instrinsic and
extrinsic)}
\usage{
is_valid_vector_borne(prob_dist, prob_dist_params, summary_stats, metadata)
}
\arguments{
\item{prob_dist}{A character string specifying the probability distribution}

\item{prob_dist_params}{A named vector of probability distribution parameters}

\item{summary_stats}{A list of summary statistics}

\item{metadata}{A list of metadata}
}
\value{
Boolean logical
}
\description{
Checks whether the data input into making an \code{epidist} class object is valid
for a vector-borne disease with two delay distributions (instrinsic and
extrinsic)
}
\examples{
# example where input is valid
is_valid_vector_borne(
  prob_dist = list("gamma", "gamma"),
  prob_dist_params = list(
    c(shape = 1, scale = 1),
    c(shape = 2, scale = 2)
  ),
  summary_stats = list(
    intrinsic = create_epidist_summary_stats(),
    extrinsic = create_epidist_summary_stats()
  ),
  metadata = create_epidist_metadata(vector_borne = TRUE)
)

# example where input is invalid
is_valid_vector_borne(
  prob_dist = list("gamma"),
  prob_dist_params = list(
    c(shape = 1, scale = 1),
    c(shape = 2, scale = 2)
  ),
  summary_stats = list(
    intrinsic = create_epidist_summary_stats(),
    extrinsic = create_epidist_summary_stats()
  ),
  metadata = create_epidist_metadata(vector_borne = TRUE)
)

is_valid_vector_borne(
  prob_dist = list("gamma", "gamma"),
  prob_dist_params = list(
    c(shape = 1, scale = 1)
  ),
  summary_stats = list(
    intrinsic = create_epidist_summary_stats(),
    extrinsic = create_epidist_summary_stats()
  ),
  metadata = create_epidist_metadata(vector_borne = TRUE)
)

is_valid_vector_borne(
  prob_dist = list("gamma", "gamma"),
  prob_dist_params = list(
    c(shape = 1, scale = 1),
    c(shape = 2, scale = 2)
  ),
  summary_stats = list(
    intrinsic = create_epidist_summary_stats(),
    extrinsic = create_epidist_summary_stats()
  ),
  metadata = create_epidist_metadata(vector_borne = FALSE)
)
}
