% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{is_vb_epidist}
\alias{is_vb_epidist}
\title{Checks whether the object is a \code{vb_epidist}}
\usage{
is_vb_epidist(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A boolean logical, \code{TRUE} if the object is an \code{vb_epidist} and
\code{FALSE} if not
}
\description{
Checks whether the object is a \code{vb_epidist}
}
\examples{
vb_edist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(transmission_mode = "vector_borne")
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)

is_vb_epidist(vb_edist)
}
