% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lines.epiparameter}
\alias{lines.epiparameter}
\title{\code{\link[=lines]{lines()}} method for \verb{<epiparameter>} class}
\usage{
\method{lines}{epiparameter}(x, cumulative = FALSE, ...)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{cumulative}{A boolean \code{logical}, default is \code{FALSE}.
\code{cumulative = TRUE} plots the cumulative distribution function (CDF).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{\link[=lines]{lines()}} method for \verb{<epiparameter>} class
}
\examples{
ebola_si <- epiparameter_db(disease = "Ebola", epi_name = "serial")
plot(ebola_si[[1]])
lines(ebola_si[[2]])
}
