% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List distributions}
\usage{
list_distributions(type = "incubation", parameters = F)
}
\arguments{
\item{type}{A \code{character} defining parameter to be listed}

\item{parameters}{A \code{logical} defining whether to show parameter values}
}
\description{
List distributions included in package
}
\examples{
list_distributions()
}
\keyword{distributions}
