% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{Lnorm}
\alias{Lnorm}
\title{Constructor for \verb{<lnorm_dist>} object}
\usage{
Lnorm(meanlog = 0, sdlog = 1, log = FALSE, ..., mean, sd)
}
\arguments{
\item{meanlog, sdlog}{mean and standard deviation of the distribution
    on the log scale with default values of \code{0} and \code{1} respectively.}
}
\value{
A \verb{<lnorm_dist>} object.
}
\description{
Constructor for \verb{<lnorm_dist>} object
}
