% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{lnorm_meansd2meanlogsdlog}
\alias{lnorm_meansd2meanlogsdlog}
\title{Converts the mean and standard deviation of a lognormal distribution into the
standard parameterisation of meanlog (mu) and sdlog (sigma)}
\usage{
lnorm_meansd2meanlogsdlog(mean, sd)
}
\arguments{
\item{mean}{Mean (expectation) of the lognormal distribution}

\item{sd}{Standard deviation of the lognormal distribution}
}
\value{
A named list with meanlog and sdlog
}
\description{
Converts the mean and standard deviation of a lognormal distribution into the
standard parameterisation of meanlog (mu) and sdlog (sigma)
}
\examples{
lnorm_meansd2meanlogsdlog(1.0, 0.4)
}
