% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{lnorm_musigma2meansd}
\alias{lnorm_musigma2meansd}
\title{Converts the mu (mean log) and sigma (standard deviation log) parameters of
the lognormal distribution to the mean and standard deviation}
\usage{
lnorm_musigma2meansd(mu, sigma)
}
\arguments{
\item{mu}{The mu parameter (mean of natural logarithm) of the lognormal
distribution}

\item{sigma}{The sigma parameter (standard deviation of the natural
logarithm) of the distribution}
}
\value{
A named list with mean and standard deviation
}
\description{
Converts the mu (mean log) and sigma (standard deviation log) parameters of
the lognormal distribution to the mean and standard deviation
}
\examples{
lnorm_musigma2meansd(1.5, 0.9)
}
