% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_json_data}
\alias{make_json_data}
\title{Creates a JSON version of the epiparameter database}
\usage{
make_json_data(
  read_path = "extdata/parameters.csv",
  write_path = "inst/extdata/data.json"
)
}
\arguments{
\item{read_path}{A relative file path from the package root directory to read
the parameters database from}

\item{write_path}{A relative file path from the package root directory to
write the JSON data file to}
}
\value{
Invisibly returns the JSON data
}
\description{
This converts the database from a data frame (read from the csv
stored in the package) into a JSON data file and saves it. This functions
allows the database to be edited and then the JSON file can be updated but
calling this function.
}
\keyword{internal}
