% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_json_schema}
\alias{make_json_schema}
\title{Creates a data schema in JSON format from a yaml file}
\usage{
make_json_schema(
  read_path = file.path("extdata", "data_dictionary.yaml"),
  write_path = file.path("inst", "extdata", "schema.json")
)
}
\arguments{
\item{read_path}{A relative file path from the package root directory to read
the yaml schema from}

\item{write_path}{A relative file path from the package root directory to
write the JSON schema to}
}
\value{
Invisibly returns the JSON schema
}
\description{
This converts the yaml data schema (data dictionary) stored in
the package into a JSON schema and saves it. This functions allows the yaml
to be the file which can be read and edited and then this file updates to
the JSON schema to match the yaml schema.
}
\keyword{internal}
