% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{mean.epidist}
\alias{mean.epidist}
\title{Mean method for \verb{<epidist>} class}
\usage{
\method{mean}{epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
A \code{numeric} mean of a distribution or \code{NA}.
}
\description{
Mean method for \verb{<epidist>} class
}
\examples{
edist <- epidist_db(
  disease = "COVID-19",
  epi_dist = "incubation_period",
  single_epidist = TRUE
)
mean(edist)
}
