% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{mean.epiparameter}
\alias{mean.epiparameter}
\title{Mean method for \verb{<epiparameter>} class}
\usage{
\method{mean}{epiparameter}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparameter>} object.}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
A \code{numeric} mean of a distribution or \code{NA}.
}
\description{
Mean method for \verb{<epiparameter>} class
}
\examples{
ep <- epiparameter_db(
  disease = "COVID-19",
  epi_dist = "incubation period",
  single_epiparameter = TRUE
)
mean(ep)
}
