% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{new_epidist}
\alias{new_epidist}
\title{Constructor for epidist class}
\usage{
new_epidist(
  disease = list(),
  epi_dist = character(),
  prob_dist = list(),
  prob_dist_params = numeric(),
  uncertainty = list(),
  summary_stats = list(),
  citation = character(),
  metadata = list(),
  method_assessment = list(),
  discretise = logical(),
  truncation = numeric(),
  notes = character(),
  auto_calc_params = TRUE
)
}
\arguments{
\item{disease}{A list contains the \verb{$disease} a character string of the
infectious disease specified in the study, and the \verb{4pathogen} a character
string or NULL of the causative agent of disease specified in study.}

\item{epi_dist}{A character string with the name of the epidemiological
distribution type}

\item{prob_dist}{A character string specifying the probability distribution}

\item{prob_dist_params}{A named vector of probability distribution parameters}

\item{uncertainty}{Either a list of named vectors with the uncertainty around
the probability distribution parameters or NULL when uncertainty around the
parameter estimates is unknown}

\item{summary_stats}{A list of summary statistics}

\item{citation}{A character string with the short form citation}

\item{metadata}{A list of metadata}

\item{method_assessment}{A list of methodological aspects assessed}

\item{discretise}{A boolean logical whether the distribution is discretised.
Default is FALSE which assumes a continuous probability distribution}

\item{truncation}{A numeric specifying the truncation point if the inferred
distribution was truncated, NULL if not or unknown.}

\item{notes}{A character string with additional notes}

\item{auto_calc_params}{}
}
\value{
epidist object
}
\description{
Create an epidist object and provides input checking. The
constructor will search whether parameters of the probability distribution
are supplied and if not look to see whether they can be inferred/extracted/
converted from summary statistics provided.
}
\examples{
epiparameter:::new_epidist(
  disease = list(
    disease = "ebola",
    pathogen = "ebola_virus"
  ),
  epi_dist = "incubation_period",
  prob_dist = list("gamma"),
  prob_dist_params = list(c(shape = 1.4, scale = 0.5)),
  uncertainty = NA,
  summary_stats = create_epidist_summary_stats(),
  metadata = create_epidist_metadata(
    sample_size = 83
  ),
  method_assessment = create_epidist_method_assessment(),
  citation = "Smith et al (2010) <10.19832/j.1366-9516.2012.09147.x>",
  notes = "No additional notes"
)
}
\keyword{internal}
