% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{new_epiparam}
\alias{new_epiparam}
\title{Constructor for \verb{<epiparam>} class}
\usage{
new_epiparam(epi_dist)
}
\arguments{
\item{epi_dist}{A \code{character} string with the name of the
epidemiological distribution type.}
}
\value{
An \verb{<epiparam>} object.
}
\description{
The constructor reads the data stored internally in the package
and subsets by epidemiological distribution (\code{epi_dist}).
}
\examples{
eparam <- epiparameter:::new_epiparam("all")
}
\keyword{internal}
