% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_tbl.R
\name{parameter_tbl}
\alias{parameter_tbl}
\title{Provide epidemiological distributions in a tabular format}
\usage{
parameter_tbl(
  multi_epidist,
  disease = "all",
  pathogen = "all",
  epi_dist = "all"
)
}
\arguments{
\item{multi_epidist}{Either an \verb{<epidist>} object or a list of \verb{<epidist>}
objects.}

\item{disease}{A \code{character} string with name of the infectious disease.}

\item{pathogen}{A \code{character} string with the name of the causative agent of
disease, or \code{NA} if not known.}

\item{epi_dist}{A \code{character} string with the name of the
epidemiological distribution type.}
}
\description{
This function subsets the epidemiological parameter library to
return only the chosen epidemiological distribution. The results are
returned as a data frame containing the disease, epidemiological
distribution, probability distribution, author of the study, and the year
of publication.
}
\details{
The \verb{<data.frame>} returned by \code{\link[=parameter_tbl]{parameter_tbl()}} contains some
vector columns and some list columns. For example, the \code{author} column is a
list as each paper can contain multiple authors.
}
\examples{
epidist_list <- epidist_db(disease = "COVID-19")
parameter_tbl(multi_epidist = epidist_list)

# example filtering an existing list to incubation periods
epidist_list <- epidist_db(disease = "COVID-19")
parameter_tbl(
  multi_epidist = epidist_list,
  epi_dist = "incubation period"
)
}
\author{
Joshua W. Lambert, Adam Kucharski
}
