% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{parameters.epidist}
\alias{parameters.epidist}
\title{Parameters method for \code{epidist} object}
\usage{
\method{parameters}{epidist}(x, ...)
}
\arguments{
\item{x}{An \code{epidist} object}

\item{...}{Additional arguments used by methods.}
}
\description{
Extracts the parameters of the distribution stored in an \code{epidist} object.
}
