% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_summary.R
\name{pathogen_summary}
\alias{pathogen_summary}
\title{Provides information on the delay distributions for a specific pathogen
stored in the \code{epiparameter} library.}
\usage{
pathogen_summary(pathogen)
}
\arguments{
\item{pathogen}{A character string with the name of the causative agent of
disease, or NULL if not known}
}
\value{
An instance of a \code{epidist_tbl} class
}
\description{
Provides information on the delay distributions for a specific pathogen
stored in the \code{epiparameter} library.
}
\examples{
\dontrun{
# summary of delay distributions available for ebola
pathogen_summary(pathogen = "ebola")

# summary of delay distributions available for SARS CoV
pathogen_summary(pathogen = "SARS_CoV")
}
}
