% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epidist}
\alias{plot.epidist}
\title{Plot method for \verb{<epidist>} class}
\usage{
\method{plot}{epidist}(x, day_range = NULL, cumulative = FALSE, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{day_range}{Either \code{NULL} or a numeric vector of length 2 with the
first and last day to plot on the x-axis.}

\item{cumulative}{A boolean \code{logical}, default is \code{FALSE}.
\code{cumulative = TRUE} plots the cumulative distribution function (CDF).}

\item{...}{\link{dots} extra arguments to plotting method, currently not used and
will produce warning if extra arguments are supplied.}
}
\description{
Plot an \verb{<epidist>} object by displaying the either the
probability density function (PDF) (in the case of continuous
distributions) or the probability mass function (PMF), (in the
case of discrete distributions) or , or the cumulative distribution
function (CDF), when \code{cumulative = TRUE}.
}
\examples{
# plot continuous epidist
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 2, scale = 1)
)

plot(edist)
plot(edist, day_range = c(5, 15))
plot(edist, cumulative = TRUE)

# plot discrete epidist
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 2, scale = 1),
  discretise = TRUE
)

plot(edist)
plot(edist, day_range = c(5, 15))
plot(edist, cumulative = TRUE)
}
\author{
Joshua W. Lambert
}
