% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.vb_epidist}
\alias{plot.vb_epidist}
\title{Plot method for \verb{<vb_epidist>} class}
\usage{
\method{plot}{vb_epidist}(x, ...)
}
\arguments{
\item{x}{A \verb{<vb_epidist>} object.}

\item{...}{\link{dots} extra arguments to be passed to either
\code{\link[=plot.epidist]{plot.epidist()}}.}
}
\description{
Plot a \verb{<vb_epidist>} object by displaying the either the
probability density function (PDF) (in the case of continuous
distributions) or the probability mass function (PMF), (in the case of
discrete distributions), or the cumulative distribution function (CDF),
for both the intrinsic and extrinsic distributions.
}
\examples{
# plot vb_epidist
dengue_dist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = FALSE
    )
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 2, scale = 2),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)

plot(dengue_dist)
}
\author{
Joshua W. Lambert
}
