% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{possible_epidist_params}
\alias{possible_epidist_params}
\title{Helper function that prints the distribution parameters accepted by
\code{epiparameter}}
\usage{
possible_epidist_params()
}
\value{
Nothing, prints to console
}
\description{
Helper function that prints the distribution parameters accepted by
\code{epiparameter}
}
\examples{
possible_epidist_params()
}
