% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{print.dist}
\alias{print.dist}
\title{This function can probably be deleted at some point as \code{\link[=.dist]{.dist()}} is internal
and there should be no point in}
\usage{
\method{print}{dist}(x, ...)
}
\description{
This function can probably be deleted at some point as \code{\link[=.dist]{.dist()}} is internal
and there should be no point in
}
