% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{print.epidist}
\alias{print.epidist}
\title{Print method for \verb{<epidist>} class}
\usage{
\method{print}{epidist}(x, header = TRUE, vb = NULL, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{header}{Boolean \code{logical} determining whether the header (first part)
of the print method is printed. This is used internally for plotting the
\verb{<vb_epidist>} class.}

\item{vb}{A \code{character} string containing whether it is the intrinsic
(\code{"Intrinsic"}) or extrinsic (\code{"Extrinsic"}) distribution for vector-borne
diseases.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
Invisibly returns an \verb{<epidist>}. Called for side-effects.
}
\description{
Print method for \verb{<epidist>} class
}
\examples{
epidist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
epidist
}
