% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{print.epiparam}
\alias{print.epiparam}
\title{Print method for epiparam class}
\usage{
\method{print}{epiparam}(x)
}
\arguments{
\item{x}{epiparam object}
}
\value{
Nothing (prints output)
}
\description{
Print method for epiparam class
}
\examples{
epiparam <- epiparam(1)
epiparam
}
