% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_db.R
\name{print.multi_epidist}
\alias{print.multi_epidist}
\title{Print method for \verb{<multi_epidist>} class}
\usage{
\method{print}{multi_epidist}(x, ...)
}
\arguments{
\item{x}{A \verb{<multi_epidist>} object.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
Invisibly returns a \verb{<multi_epidist>}. Called for side-effects.
}
\description{
Print method for \verb{<multi_epidist>} class
}
