% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_f.R
\name{prob_f}
\alias{prob_f}
\title{Parameter probability mass function}
\usage{
prob_f(pathogen, type, study = NULL, pmf = T)
}
\arguments{
\item{pathogen}{A \code{character} specifying type of parameter: "incubation" or "onset_to_admission"}

\item{study}{A \code{character} specifying dataset to use. Defaults to study with largest sample size.}

\item{pmf}{A \code{logical} specifying whether to output probability mass function (\code{TRUE}) or cumulative mass function (\code{FALSE}). Defaults to \code{TRUE}.}
}
\description{
Parameter probability distribution by day
}
\examples{
prob_f()
}
\keyword{incubation}
