% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{[.epiparam}
\alias{[.epiparam}
\title{Subsetting method for \code{epiparam}}
\usage{
\method{[}{epiparam}(epiparam, ...)
}
\arguments{
\item{epiparam}{An \code{epiparam} object.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
An \code{epiparam} object or a \code{data.frame}
}
\description{
If the subsetting invalidates the \code{epiparam} object (defined by
its invariants, and encoded in \code{\link[=validate_epiparam]{validate_epiparam()}}) the subsetting will
return a data frame with a message to console stating the class of the object
has been converted to \code{data.frame} with the other attributes of the class
preserved.
}
