% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{summary.epiparam}
\alias{summary.epiparam}
\title{Summary method for \verb{<epiparam>} class}
\usage{
\method{summary}{epiparam}(object, ...)
}
\arguments{
\item{object}{An \verb{<epiparam>} object.}

\item{...}{\link{dots} not used, extra arguments supplied will cause a warning.}
}
\value{
data frame of information
}
\description{
Summary method for \verb{<epiparam>} class
}
\examples{
x <- epiparam()
summary(x)
}
