% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{tail}
\alias{tail}
\title{tail method for epiparam class}
\usage{
tail(x)
}
\arguments{
\item{x}{epiparam object}
}
\value{
Data frame
}
\description{
tail method for epiparam class
}
\examples{
tail(epiparam())
}
