% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{test_epiparameter}
\alias{test_epiparameter}
\title{Test whether an object is a valid \verb{<epiparameter>} object}
\usage{
test_epiparameter(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A boolean \code{logical} whether the object is a valid \verb{<epiparameter>}
object (prints message when invalid \verb{<epiparameter>} object is provided).
}
\description{
Test whether an object is a valid \verb{<epiparameter>} object
}
