% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_epiparameter}
\alias{theme_epiparameter}
\title{A custom \pkg{ggplot2} theme for \pkg{epiparameter}}
\usage{
theme_epiparameter(...)
}
\arguments{
\item{...}{\link{Dots} Extra arguments to modify \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.}
}
\value{
A \pkg{ggplot2} \link[ggplot2:theme]{ggplot2::theme} object.
}
\description{
A custom \pkg{ggplot2} theme for \pkg{epiparameter}
}
