% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{validate_epidist}
\alias{validate_epidist}
\title{\code{epidist} class validator}
\usage{
validate_epidist(epidist)
}
\arguments{
\item{epidist}{An \code{epidist} object}
}
\value{
Invisibly returns an \code{\link{epidist}}. Called for side-effects (errors
when invalid \code{epidist} object is provided).

Nothing, errors when invalid \code{epidist} object is provided
}
\description{
\code{epidist} class validator
}
