% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{validate_epiparam}
\alias{validate_epiparam}
\title{Validator for \verb{<epiparam>} class}
\usage{
validate_epiparam(epiparam, reconstruct = FALSE)
}
\arguments{
\item{epiparam}{An \verb{<epiparam>} object.}

\item{reconstruct}{A boolean \code{logical} determining whether the validation
should be class specific. When \code{TRUE} the input object must be of type
\verb{<epiparam>} (default), when \code{FALSE} the input object can be of another
class, e.g. data frame. This argument is used in reconstruction operations
see \code{\link[=epiparam_reconstruct]{epiparam_reconstruct()}}.}
}
\value{
Invisibly returns an \verb{<epiparam>}. Called for side-effects
(errors when invalid \verb{<epiparam>} object is provided).
}
\description{
Validator for \verb{<epiparam>} class
}
