% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter.R
\name{validate_epiparameter}
\alias{validate_epiparameter}
\title{Validator for \verb{<epiparameter>} class}
\usage{
validate_epiparameter(epiparameter)
}
\arguments{
\item{epiparameter}{An \verb{<epiparameter>} object}
}
\value{
Invisibly returns an \verb{<epiparameter>}. Called for side-effects
(errors when invalid \verb{<epiparameter>} object is provided).
}
\description{
Validator for \verb{<epiparameter>} class
}
