% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{validate_vb_epidist}
\alias{validate_vb_epidist}
\title{Validator for \verb{<vb_epidist>} class}
\usage{
validate_vb_epidist(vb_epidist)
}
\arguments{
\item{vb_epidist}{A \verb{<vb_epidist>} object}
}
\value{
Invisibly returns a \verb{<vb_epidist>}. Called for side-effects (errors
when invalid \verb{<vb_epidist>} object is provided).
}
\description{
Validator for \verb{<vb_epidist>} class
}
