% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{vb_epidist}
\alias{vb_epidist}
\title{Create a \verb{<vb_epidist>} object}
\usage{
vb_epidist(intrinsic_epidist, extrinsic_epidist)
}
\arguments{
\item{intrinsic_epidist}{An \verb{<epidist>} object.}

\item{extrinsic_epidist}{An \verb{<epidist>} object.}
}
\value{
\verb{<vb_epidist>} object
}
\description{
The \verb{<vb_epidist>} class is an extension of the
\verb{<epidist>} class (although not a subclass of \verb{<epidist>}). It is is
used to store epidemiological parameters for vector-borne diseases.
It has the same methods (\code{print()}, \code{format()}, \code{plot()}, \code{generate()},
\code{cdf()}, \code{density()}, \code{quantile()}) as the \verb{<epidist>} class and
therefore they are used identically.
}
\details{
The \verb{<epidist>} objects should contain metadata
(\code{epidist$metadata}) indicating it is a vector-borne disease
(\code{epidist$metadata$transmission_mode = "vector_borne"}) and the extrinsic
distribution should indicate in the metadata that it is the extrinsic
distribution (\code{epidist$metadata$extrinsic = TRUE}). If these two aspects
are not given the construction of the class will throw a warning.
}
\examples{
vb <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    pathogen = "dengue_virus",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = FALSE
    )
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    pathogen = "dengue_virus",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 2, scale = 2),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)
}
\seealso{
\code{\link[=epidist]{epidist()}}
}
