% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{weibull_meansd2shapescale}
\alias{weibull_meansd2shapescale}
\title{Converts the mean and standard deviation of the weibull distribution to the
shape and scale parameterisation.}
\usage{
weibull_meansd2shapescale(mean, sd)
}
\arguments{
\item{mean}{The mean of the weibull distribution}

\item{sd}{The standard deviation of the weibull distribution}
}
\value{
A named list with shape and scale parameters
}
\description{
Converts the mean and standard deviation of the weibull distribution to the
shape and scale parameterisation.
}
\examples{
weibull_meansd2shapescale(mean = 1.5, sd = 0.5)
}
