% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{weibull_shapescale2meansd}
\alias{weibull_shapescale2meansd}
\title{Converts the shape and scale parameters of the weibull distribution to the
mean and standard deviation}
\usage{
weibull_shapescale2meansd(shape, scale)
}
\arguments{
\item{shape}{A single numeric of the shape parameter of the weibull
distribution}

\item{scale}{A single numeric of the scale parameter of the weibull
distribution}
}
\value{
A named list with mean and standard deviation
}
\description{
Converts the shape and scale parameters of the weibull distribution to the
mean and standard deviation
}
\examples{
weibull_shapescale2meansd(shape = 2, scale = 1)
}
