% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.multi_epiparameter}
\alias{aggregate.multi_epiparameter}
\title{Convert a \verb{<multi_epiparameter>} object to a single \verb{<epiparameter>}}
\usage{
\method{aggregate}{multi_epiparameter}(
  x,
  synthesis_method = c("meta_analysis", "naive"),
  weights = c("sample_size", "unweighted"),
  allowed_pairs = list(c("mean", "sd"), c("median", "sd"), c("median", "IQR")),
  ...
)
}
\arguments{
\item{x}{A \verb{<multi_epiparameter>} object.}

\item{synthesis_method}{The method used to aggregate the estimates. Can be
either \code{"meta_analysis"} (default) or \code{"naive"}.}

\item{weights}{The method used to weight the estimates. Can be either
\code{"sample_size"} (default) or \code{"unweighted"}.}

\item{allowed_pairs}{A list of pairs of population-level summary statistics.
to}

\item{...}{\link{dots} Not used, extra arguments supplied will cause a warning.}
}
\value{
An \verb{<epiparameter>} object with synthetised central estimate value
and uncertainty
}
\description{
Convert a \verb{<multi_epiparameter>} object to a single \verb{<epiparameter>}
}
\examples{
epiparameter_db(epi_dist = "serial interval", disease = "ebola")

}
