% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{as_epidist}
\alias{as_epidist}
\title{Convert \verb{<epiparam>} to \verb{<epidist>}}
\usage{
as_epidist(x)
}
\arguments{
\item{x}{A \verb{<epiparam>} object.}
}
\value{
An \verb{<epidist>} object or a list of \verb{<epidist>} objects.
}
\description{
Convert the entries (rows) of an \verb{<epiparam>} object to one or
list of several \verb{<epidist>} objects.

Epidemiological distributions and parameters can be converted from database
entries (i.e. rows in \verb{<epiparam>}) into \verb{<epidist>} objects in order
use the distribution functions (see \code{?epidist_distribution_functions}) and
and other methods of the \verb{<epidist>} class.
}
\examples{
\donttest{
eparam <- epiparam()
as_epidist(eparam[1, ])
}
}
