% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist}
\alias{as_epidist}
\title{Convert to an \verb{<epidist>} object}
\usage{
as_epidist(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
An \verb{<epidist>} object or list of \verb{<epidist>} objects.
}
\description{
Convert from an \R object to an \verb{<epidist>} object. If conversion is not
possible the function will error.
}
