% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_epidist}
\alias{as_epidist}
\title{Convert to an \verb{<epidist>} object}
\usage{
as_epidist(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
An \verb{<epidist>} object or list of \verb{<epidist>} objects.
}
\description{
Convert from an \R object to an \verb{<epidist>} object. If conversion is not
possible the function will error.
}
\details{
To create the full citation the information from the article table from the
\pkg{epireview} package of the corresponding entry will need to be passed
to function via the \code{...} argument. The argument should be called \code{article},
as it will be matched by name by \code{$}.
}
