% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{as_epidist}
\alias{as_epidist}
\title{Converts rows of an epiparam object into an epidist object}
\usage{
as_epidist(x)
}
\arguments{
\item{x}{A epiparam object}
}
\value{
An epidist object or a list of epidist objects
}
\description{
Can convert the entries (rows) of an epiparam object to one or
list of several epidist objects. This allows subsetting of epiparam records
into a class that can then be used to to calculate the cdf, random number
generation and other methods of the epidist class
(see \code{?epidist_distribution_functions}).
}
\examples{
eparam <- epiparam()
as_epidist(eparam[1, ])
}
