% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_epiparam.R
\name{bind_epiparam}
\alias{bind_epiparam}
\title{Binds any epi data class in \pkg{epiparameter} (\code{\link{epidist}}, \code{\link{vb_epidist}},
\code{\link{epiparam}}) or data frame to an \code{\link{epiparam}} object.}
\usage{
bind_epiparam(epiparam, epi_obj)
}
\arguments{
\item{epiparam}{An \code{epiparam} object}

\item{epi_obj}{Either an \code{epidist}, \code{vb_epidist}, \code{epiparam} or list of
\code{epidist} objects. It can also be a data frame as long as the columns conform
to the columsn of an \code{epiparam} object.}
}
\value{
An \code{epiparam} object
}
\description{
Binds any epi data class in \pkg{epiparameter} (\code{\link{epidist}}, \code{\link{vb_epidist}},
\code{\link{epiparam}}) or data frame to an \code{\link{epiparam}} object.
}
\details{
The \code{\link{epiparam}} class holds the library of epidemiological
parameters that is stored in the \pkg{epiparameter} R package and can be
manipulated. The \code{\link[=bind_epiparam]{bind_epiparam()}} function allows users to add entries to
the library by binding them to the bottom of an existing \code{\link{epiparam}} object
loaded in R.
}
\examples{
eparam <- epiparam()
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "lnorm",
  prob_distribution_params = c(meanlog = 1, sdlog = 1)
)
bind_epiparam(eparam, edist)
}
