% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{clean_disease}
\alias{clean_disease}
\title{Standardise the names of diseases}
\usage{
clean_disease(x)
}
\arguments{
\item{x}{A \code{character} string specifying a disease.}
}
\value{
A \code{character} vector of equal length to the input.
}
\description{
Standardise the names of diseases
}
