% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_geom_params}
\alias{convert_geom_params}
\title{Converts the parameters of the geometric distribution to summary statistics}
\usage{
convert_geom_params(prob)
}
\arguments{
\item{prob}{The probability parameter (p) of the geometric.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and kurtosis.
}
\description{
Converts the probability (\code{prob}) of the geometric distribution
to a number of summary statistics which can be calculated analytically given
the geometric parameter.
}
\examples{
convert_geom_params(prob = 1)
}
