% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_geom_summary_stats}
\alias{convert_geom_summary_stats}
\title{Converts the summary statistics to parameters of the geometric distribution}
\usage{
convert_geom_summary_stats(...)
}
\arguments{
\item{...}{Summary statistics to be used for calculating the geometric
distribution parameters. All arguments must be correctly named, see details
for possible input.}
}
\value{
A list of one element, the probability parameter
}
\description{
Converts the summary statistics of the geometric
distribution the parameter (\code{prob}) of the geometric distribution.
}
\details{
The arguments input are captured and handled by name. Therefore,
the names of the arguments must be correct (case-sensitive). Possible input
summary statistics are: \code{mean}, \code{median}, \code{mode}, \code{var}, \code{sd}, \code{cv},
\code{skewness}, \code{kurtosis}.

Not every combination of input statistics is a viable conversion, and may
error. Common conversions that are implemented are mean and any measure of
spread (\code{var}, \code{sd} and \code{cv}), or \code{median} and \code{sd}.
}
\examples{
convert_geom_summary_stats(mean = 3)
}
