% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist_params.R
\name{convert_params}
\alias{convert_params}
\title{Converts summary statistics to parameters of a probability distribution}
\usage{
convert_params(summary_stats, prob_dist)
}
\arguments{
\item{summary_stats}{A list of summary statistics, use
\code{create_epidist_summary_stats()} to create list.}

\item{prob_dist}{A character string specifying the probability
distribution}
}
\value{
Named vector
}
\description{
The summary_stats object needs to have a class attribute with
the name of the probability distribution in order for the correct conversion
function to be dispatched.
}
\keyword{internal}
