% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_params.R
\name{convert_params_to_summary_stats}
\alias{convert_params_to_summary_stats}
\title{Convert the parameter(s) of a distribution to summary statistics}
\usage{
convert_params_to_summary_stats(
  distribution = c("lnorm", "gamma", "weibull", "nbinom", "geom"),
  ...
)
}
\arguments{
\item{distribution}{A \code{character} specifying distribution to use.
Default is \code{lnorm}; also takes \code{gamma} and \code{weibull}, \code{nbinom} and \code{geom}.}

\item{...}{\code{Numeric} named parameter(s) used to convert to summary
statistics. An example is the meanlog and sdlog parameters of the
lognormal (\code{lnorm}) distribution.}
}
\value{
A list of eight elements including: mean, median, mode,
variance (\code{var}), standard deviation (\code{sd}), coefficient of variation (\code{cv}),
skewness, and excess kurtosis (\code{ex_kurtosis}).
}
\description{
Converts the parameters for a range of distributions to a
number of summary statistics. All summary statistics are calculated
analytically given the parameters.
}
\details{
The distribution names and parameter names follow the style of
distributions in R, for example the lognormal distribution is \code{lnorm},
and its parameters are \code{meanlog} and \code{sdlog}.
}
\examples{
convert_params_to_summary_stats(
  distribution = "lnorm", meanlog = 1, sdlog = 2
)
convert_params_to_summary_stats(
  distribution = "gamma", shape = 1, scale = 1
)
convert_params_to_summary_stats(
  distribution = "nbinom", prob = 0.5, dispersion = 2
)
}
