% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_citation}
\alias{create_epidist_citation}
\title{A helper function when creating an epidist object to create a summary
statistics list with sensible defaults, type checking and arguments to help
remember which summary statistics can be accepted in the list
A helper function when creating an epidist object to create a citation list
with sensible defaults, type checking and arguments to help remember which
citation information is accepted in the list.}
\usage{
create_epidist_citation(
  author = NA_character_,
  year = NA_integer_,
  title = NA_character_,
  journal = NA_character_,
  DOI = NA_character_,
  PMID = NA_character_
)
}
\arguments{
\item{author}{A \code{character} string of the surname of the first author. This
can be underscore separated from a second author, or underscore separated
from "etal" if there are more than two authors.}

\item{year}{A \code{numeric} of the year of publication}

\item{title}{A \code{character} string with the title of the article that
published the epidemiological parameters.}

\item{journal}{A \code{character} string with the name of the journal that
published the article that published the epidemiological parameters.
This can also be a pre-print server, e.g., medRxiv.}

\item{DOI}{A \code{character} string of the Digital Object Identifier (DOI)
assigned to papers which are unique to each paper.}

\item{PMID}{A \code{character} string with the PubMed unique identifier number
assigned to papers to give them a unique identifier within PubMed.}
}
\value{
A character string of the formatted short citation
}
\description{
A helper function when creating an epidist object to create a summary
statistics list with sensible defaults, type checking and arguments to help
remember which summary statistics can be accepted in the list
A helper function when creating an epidist object to create a citation list
with sensible defaults, type checking and arguments to help remember which
citation information is accepted in the list.
}
\examples{
create_epidist_citation(
  author = "Smith_etal",
  year = 2002,
  DOI = "10.19832/j.1366-9516.2012.09147.x"
)
}
