% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{create_epidist_uncertainty}
\alias{create_epidist_uncertainty}
\title{A helper function when creating uncertainty for the parameters of the
distribution for the epidist object}
\usage{
create_epidist_uncertainty(ci, ci_interval, ci_type)
}
\arguments{
\item{ci}{A numeric vector of length two with the lower and upper bound of
the confidence interval or credible interval}

\item{ci_interval}{A numeric specifying the interval for the ci, e.g. 95 is
95\% ci}

\item{ci_type}{A character string, either "confidence interval" or "credible
interval"}
}
\value{
Numeric vector with an attribute
}
\description{
A helper function when creating uncertainty for the parameters of the
distribution for the epidist object
}
\examples{
a = 1
}
