% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{epidist}
\alias{epidist}
\title{Create an \code{epidist} object}
\usage{
epidist(
  disease,
  pathogen = NULL,
  epi_distribution,
  prob_distribution = NULL,
  prob_distribution_params = NULL,
  uncertainty = create_epidist_uncertainty(),
  summary_stats = create_epidist_summary_stats(),
  citation = create_epidist_citation(),
  metadata = create_epidist_metadata(),
  method_assessment = create_epidist_method_assessment(),
  discretise = FALSE,
  truncation = NULL,
  notes = NULL
)
}
\arguments{
\item{disease}{A character string with name of the infectious disease}

\item{pathogen}{A character string with the name of the causative agent of
disease, or NULL if not known}

\item{epi_distribution}{A character string with the name of the
epidemiological distribution type}

\item{prob_distribution}{A character string specifying the probability
distribution}

\item{prob_distribution_params}{A named vector of probability distribution
parameters}

\item{uncertainty}{Either a list of named vectors with the uncertainty around
the probability distribution parameters or NULL when uncertainty around the
parameter estimates is unknown}

\item{summary_stats}{A list of summary statistics}

\item{citation}{A character string with the citation of the source of the
data or the paper that inferred the distribution parameters}

\item{metadata}{A list of metadata, this can include: sample size, whether
the disease is vector-borne, etc. It is assumed that the disease is not
vector-borne and that the distribution is intrinsic (e.g. not an extrinsic
delay distribution such as extrinsic incubation period) unless
\code{vector_borne = TRUE} is contained in the metadata}

\item{method_assessment}{A list of methodological aspects assessed}

\item{discretise}{A boolean logical whether the distribution is discretised.
Default is FALSE which assumes a continuous probability distribution}

\item{truncation}{A numeric specifying the truncation point if the inferred
distribution was truncated, NULL if not or unknown.}

\item{notes}{A character string with any additional information about the
data, inference method or disease.}
}
\value{
An \code{epidist} object
}
\description{
The \code{epidist} class is used to store epidemiological parameters
for a single disease. These epidemiological parameters cover a variety of
aspects including delay distributions (e.g. incubation periods and serial
intervals, among others) and offspring distributions.
}
\examples{
ebola_incubation <- epidist(
  disease = "ebola",
  pathogen = "ebola_virus",
  epi_distribution = "incubation",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1),
  uncertainty = NULL,
  summary_stats = create_epidist_summary_stats(
    mean = 2,
    sd = 1
  ),
  citation = create_epidist_citation(
    author = "Smith_etal",
    year = 2010,
    DOI = "10.19832/j.1366-9516.2012.09147.x"
  ),
  metadata = create_epidist_metadata(
    sample_size = 10,
    region = "UK",
    vector_borne = FALSE,
    inference_method = "MLE"
  ),
  method_assessment = create_epidist_method_assessment(
    censorred = TRUE
  ),
  discretise = FALSE,
  truncation = NULL,
  notes = "No notes"
)
}
