% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter_db.R
\name{epidist_db}
\alias{epidist_db}
\title{Create \verb{<epiparameter>} object(s) directly from the epiparameter library
(database)}
\usage{
epidist_db(
  disease = "all",
  pathogen = "all",
  epi_dist = "all",
  author = NULL,
  subset = NULL,
  single_epiparameter = FALSE
)
}
\arguments{
\item{disease}{A \code{character} string specifying the disease.}

\item{pathogen}{A \code{character} string specifying the pathogen.}

\item{epi_dist}{A \code{character} string specifying the epidemiological
distribution. See details for full list of epidemiological distributions.}

\item{author}{A \code{character} string specifying the author of the study
reporting the distribution. Only the first author will be matched. It is
recommended to use the family name as first names may or may not be
initialised.}

\item{subset}{Either \code{NULL} or a valid R expressions that evaluates to
logicals to subset the list of \verb{<epiparameter>}, or a function that can be
applied over a list of \verb{<epiparameter>} objects.

Subsetting (using \code{subset}) can be combined with the subsetting done with
the \code{disease} and \code{epi_dist} arguments (and \code{author} if specified). If left
as \code{NULL} (default) no subsetting is carried out.

The \code{subset} argument is similar to subsetting a \verb{<data.frame>}, but the
difference is that fixed comparisons and not vectorised comparisons are
needed. For example \code{sample_size > 10} is a valid subset expression, but
\code{sample_size == max(sample_size)}, which would be a valid subset expression
for a \verb{<data.frame>} does not work. The vectorised expression will often
not error, but will likely return unexpected results. For the
\code{sample_size == max(sample_size)} example it will always return \code{TRUE}
(except for \code{NA}s) as it is a single numeric so will be equal to it's
max value.

The expression should be specified without using the data object name
(e.g. \code{df$var}) and instead just \code{var} should be supplied. In
other words, this argument uses non-standard evaluation, just as the
\code{subset} argument in \code{\link[=subset]{subset()}}, and is similar to \verb{<data-masking>} used
by the \code{dplyr} package.}

\item{single_epiparameter}{A boolean \code{logical} determining whether a single
\verb{<epiparameter>} or multiple entries from the library can be returned if
matched by the other arguments (\code{disease}, \code{epi_dist}, \code{author}). This
argument is used to prevent multiple sets of parameters being returned
when only one is wanted.

\strong{Note}: If multiple entries match the arguments supplied and
\code{single_epiparameter = TRUE} then the \verb{<epiparameter>} that is parameterised
(and accounts for truncation if available) and has the largest sample size
will be returned (see \code{\link[=is_parameterised]{is_parameterised()}}). If multiple entries are equal
after this sorting the first entry will be returned.}
}
\value{
An \verb{<epiparameter>} object or list of \verb{<epiparameter>} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{epidist_db()} has been renamed \code{\link[=epiparameter_db]{epiparameter_db()}}. Please use
\code{epiparameter_db()} instead as the \code{epidist_db()} alias will be removed from
the package in the future.
}
\keyword{internal}
