% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_db.R
\name{epidist_db}
\alias{epidist_db}
\title{Create \verb{<epidist>} object(s) directly from the epiparameter library
(database)}
\usage{
epidist_db(
  disease = "all",
  epi_dist = "all",
  author = NULL,
  subset = NULL,
  single_epidist = FALSE
)
}
\arguments{
\item{disease}{A \code{character} string specifying the disease.}

\item{epi_dist}{A \code{character} string specifying the epidemiological
distribution. See details for full list of epidemiological distributions.}

\item{author}{A \code{character} string specifying the author of the study
reporting the distribution. Only the first author will be matched. It is
recommended to use the family name as first names may or may not be
initialised.}

\item{subset}{Either \code{NULL} or a valid R expressions that evaluates to
logicals to subset the list of \verb{<epidist>}, or a function that can be
applied over a list of \verb{<epidist>} objects.

Subsetting (using \code{subset}) can be combined with the subsetting done with
the \code{disease} and \code{epidist} arguments (and \code{author} if specified). If left
as \code{NULL} (default) no subsetting is carried out.

The \code{subset} argument is similar to subsetting a \verb{<data.frame>}, but the
difference is that fixed comparisons and not vectorised comparisons are
needed. For example \code{sample_size > 10} is a valid subset expression, but
\code{sample_size == max(sample_size)}, which would be a valid subset expression
for a \verb{<data.frame>} does not work. The vectorised expression will often
not error, but will likely return unexpected results. For the
\code{sample_size == max(sample_size)} example it will always return \code{TRUE}
(except for \code{NA}s) as it is a single numeric so will be equal to it's
max value.

The expression should be specified without using the data object name
(e.g. \code{df$var}) and instead just \code{var} should be supplied. In
other words, this argument uses non-standard evaluation, just as the
\code{subset} argument in \code{\link[=subset]{subset()}}, and is similar to \verb{<data-masking>} used
by the \code{dplyr} package.}

\item{single_epidist}{A boolean \code{logical} determining whether a single
\verb{<epidist>} or multiple entries from the library can be returned if
matched by the other arguments (\code{disease}, \code{epi_dist}, \code{author}). This
argument is used to prevent multiple sets of parameters being returned
when only one is wanted.

\strong{Note}: If multiple entries match the arguments supplied and
\code{single_epidist = TRUE} then the \verb{<epidist>} that is parameterised and
has the largest sample size will be returned (see \code{\link[=is_parameterised]{is_parameterised()}}).
If multiple entries are equal after this sorting the first entry will
be returned.}
}
\value{
An \verb{<epidist>} object or list of \verb{<epidist>} objects.
}
\description{
Extract \verb{<epidist>} object(s) directly from
the library of epidemiological parameters. The \package{epiparameter}
library of epidemiological parameters is compiled from primary literature
sources. The list output from \code{\link[=epidist_db]{epidist_db()}} can be subset by the data it
contains, for example by: disease, pathogen, epidemiological distribution,
sample size, region, etc.

If a distribution from a specific study is required, the \code{author} argument
can be specified.

Multiple entries (\verb{<epidist>} objects) can be returned, use the arguments
to subset entries and use \code{single_epidist = TRUE} to force a single
\verb{<epidist>} to be returned.
}
\details{
\code{disease}, \code{epi_dist} and \code{author} are given as individual arguments
as these are the most common variables to subset the parameter library by.
The \code{subset} argument facilitates all other subsetting of rows to select the
\verb{<epidist>} object(s) desired. To subset based on multiple variables separate
each expression with \code{&}.

List of epidemiological distributions:
\itemize{
\item "all" (default, returns all entries in library)
\item "incubation period"
\item "onset to hospitalisation"
\item "onset to death"
\item "serial interval"
\item "generation time"
\item "offspring distribution"
\item "hospitalisation to death"
\item "hospitalisation to discharge"
\item "notification to death"
\item "notification to discharge"
\item "onset to discharge"
\item "onset to ventilation"
}
}
\examples{
epidist_db(disease = "influenza", epi_dist = "serial_interval")

# example using custom subsetting
eparam <- epidist_db(
  disease = "SARS",
  epi_dist = "offspring_distribution",
  subset = sample_size > 40
)

# example using functional subsetting
eparam <- epidist_db(
  disease = "COVID-19",
  epi_dist = "incubation_period",
  subset = is_parameterised
)

# example forcing a single <epidist> to be returned
eparam <- epidist_db(
  disease = "SARS",
  epi_dist = "offspring_distribution",
  single_epidist = TRUE
)
}
