% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{epidist_distribution_functions}
\alias{epidist_distribution_functions}
\alias{density.epidist}
\alias{cdf.epidist}
\alias{quantile.epidist}
\alias{generate.epidist}
\title{PDF, CDF, PMF, quantiles and random number generation for \verb{<epidist>} objects}
\usage{
\method{density}{epidist}(x, at, ...)

\method{cdf}{epidist}(x, q, ..., log = FALSE)

\method{quantile}{epidist}(x, p, ...)

\method{generate}{epidist}(x, times, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{at}{The quantiles to evaluate at.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}

\item{q}{The quantiles to evaluate at.}

\item{log}{If \code{TRUE}, probabilities will be given as log probabilities.}

\item{p}{The probabilities to evaluate at.}

\item{times}{The number of random samples.}
}
\value{
\code{numeric} vector.
}
\description{
The \verb{<epidist>} object holds a probability distribution which
can either be a continuous or discrete distribution. These are the density,
cumulative distribution, quantile and random number generation functions.
These operate on any distribution that can be included in an \verb{<epidist>}
object.
}
\examples{
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# example of each distribution method for an `epidist` object
stats::density(edist, at = 1)
distributional::cdf(edist, q = 1)
stats::quantile(edist, p = 0.2)
distributional::generate(edist, times = 10)
}
\keyword{epidist_distribution_functions}
