% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R, R/vb_epidist.R
\name{epidist_distribution_functions}
\alias{epidist_distribution_functions}
\alias{density.epidist}
\alias{cdf.epidist}
\alias{quantile.epidist}
\alias{generate.epidist}
\alias{density.vb_epidist}
\alias{cdf.vb_epidist}
\alias{quantile.vb_epidist}
\alias{generate.vb_epidist}
\title{PDF, CDF, PMF, quantiles and random number generation for \verb{<epidist>} and
\verb{<vb_epidist>} objects}
\usage{
\method{density}{epidist}(x, at, ...)

\method{cdf}{epidist}(x, q, ...)

\method{quantile}{epidist}(x, p, ...)

\method{generate}{epidist}(x, times, ...)

\method{density}{vb_epidist}(x, at, ...)

\method{cdf}{vb_epidist}(x, q, ...)

\method{quantile}{vb_epidist}(x, p, ...)

\method{generate}{vb_epidist}(x, times, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} or \verb{<vb_epidist>} object.}

\item{at}{The quantiles to evaluate at.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}

\item{q}{The quantiles to evaluate at.}

\item{p}{The probabilities to evaluate at.}

\item{times}{The number of random samples.}
}
\value{
If an \verb{<epidist>} object is given a numeric vector is returned, if an
\verb{<vb_epidist>} object is given a list of two elements each with a numeric
vector is returned.
}
\description{
The \verb{<epidist>} object holds a probability distribution which
can either be a continuous or discrete distribution. These are the density,
cumulative distribution, quantile and random number generation functions.
These operate on any distribution that can be included in an \verb{<epidist>}
object.
}
\examples{
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

# example of each distribution method for an `epidist` object
stats::density(edist, at = 1)
distributional::cdf(edist, q = 1)
stats::quantile(edist, p = 0.2)
distributional::generate(edist, times = 10)

vb_edist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(transmission_mode = "vector_borne")
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)

# example of each distribution method for an `vb_epidist` object
stats::density(vb_edist, at = 1)
distributional::cdf(vb_edist, q = 1)
stats::quantile(vb_edist, p = 0.2)
distributional::generate(vb_edist, times = 10)
}
\keyword{epidist_distribution_functions}
