% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R, R/vb_epidist.R
\name{density.epidist}
\alias{density.epidist}
\alias{cdf.epidist}
\alias{quantile.epidist}
\alias{generate.epidist}
\alias{density.vb_epidist}
\alias{cdf.vb_epidist}
\alias{quantile.vb_epidist}
\alias{generate.vb_epidist}
\title{PDF, CDF, PMF, quantiles and random number generation for \code{epidist} and
\code{vb_epidist} objects}
\usage{
\method{density}{epidist}(x, at, ...)

\method{cdf}{epidist}(x, q, ...)

\method{quantile}{epidist}(x, p, ...)

\method{generate}{epidist}(x, times, ...)

\method{density}{vb_epidist}(x, at, ...)

\method{cdf}{vb_epidist}(x, q, ...)

\method{quantile}{vb_epidist}(x, p, ...)

\method{generate}{vb_epidist}(x, times, ...)

\method{density}{vb_epidist}(x, at, ...)

\method{cdf}{vb_epidist}(x, q, ...)

\method{quantile}{vb_epidist}(x, p, ...)

\method{generate}{vb_epidist}(x, times, ...)
}
\arguments{
\item{x}{An epidist or vb_epidist object}

\item{at}{The quantiles to evaluate at}

\item{q}{The quantiles to evaluate at}

\item{p}{The probabilities to evaluate at}

\item{time}{The number of random samples}
}
\value{
Numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector

List of two elements each with a numeric vector
}
\description{
The epidist object holds a probability distribution which can
either be a continuous or discrete distribution. These are the density,
cumulative distribution, quantile and random number generation functions.
These operate on any distribution that can be included in an epidist object.
}
