% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{density.epidist}
\alias{density.epidist}
\alias{cdf.epidist}
\alias{quantile.epidist}
\alias{generate.epidist}
\title{PDF, CDF, PMF, quantiles and random number generation for epidist objects}
\usage{
\method{density}{epidist}(x, at, ...)

\method{cdf}{epidist}(x, q, ...)

\method{quantile}{epidist}(x, p, ...)

\method{generate}{epidist}(x, times, ...)
}
\arguments{
\item{x}{An epidist object}

\item{at}{The quantiles to evaluate at}

\item{q}{The quantiles to evaluate at}

\item{p}{The probabilities to evaluate at}

\item{time}{The number of random samples}
}
\value{
Numeric vector
}
\description{
The epidist object holds a probability distribution which can
either be a continuous or discrete distribution. These are the density,
cumulative distribution, quantile and random number generation functions.
These operate on any distribution that can be included in an epidist object.
}
