% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{epiparam}
\alias{epiparam}
\title{Create an \code{epiparam} object}
\usage{
epiparam(
  epi_dist = c("all", "incubation_period", "onset_to_admission", "onset_to_death",
    "serial_interval", "generation_time")
)
}
\arguments{
\item{epi_dist}{A character string of which epidemiological distributions
to select}
}
\value{
\code{epiparam} object
}
\description{
Create an \code{epiparam} object
}
\examples{
# the object can be made without arguments
eparam <- epiparam()

# specifying incubation periods
incub_eparam <- epiparam("incubation")
}
