% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparameter-package.R
\docType{package}
\name{epiparameter-package}
\alias{epiparameter-package}
\title{epiparameter: Library of Epidemiological Parameters with Helper Functions and Classes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Library of epidemiological parameters for infectious diseases extracted from the literature, and a set of classes and helper functions for working with parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/epiparameter/}
  \item \url{https://epiverse-trace.github.io/epiparameter/}
  \item Report bugs at \url{https://github.com/epiverse-trace/epiparameter/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [copyright holder]
  \item Carmen Tamayo \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID})
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor, reviewer]
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID}) [contributor]
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [reviewer]
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [reviewer]
  \item Chris Hartgerink \email{chris@data.org} (\href{https://orcid.org/0000-0003-1050-6809}{ORCID}) [reviewer]
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [reviewer]
}

}
\keyword{internal}
