% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{epiparameter_options}
\alias{epiparameter_options}
\title{Package options}
\format{
An object of class \code{list} of length 2.
}
\usage{
epiparameter_options
}
\description{
Options to modify the printing of \pkg{epiparameter} objects.
Currently options are only used to modify the printing of the
\verb{<multi_epidist>} class.
}
\details{
Options are set with \code{\link[=options]{options()}} and retrieved with \code{\link[=getOption]{getOption()}}.

If options are changed the \pkg{epiparameter} package will need to be
reloaded for new options to be taken into account. Options can be set in
the \code{.Rprofile} to persist across R sessions.
}
\keyword{datasets}
