% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epireview_core_cols}
\alias{epireview_core_cols}
\title{A vector of \code{character} strings with the core column names of the
epidemiological parameter data exported by the \pkg{epireview} R package.}
\format{
\subsection{\code{epireview_core_cols}}{

A \code{character} vector with 58
elements.

The data is taken as the intersection of the column names of each of the
disease parameter tables in the \pkg{epireview} R package.
}
}
\source{
\url{https://github.com/mrc-ide/epireview}
}
\usage{
epireview_core_cols
}
\description{
A vector of \code{character} strings with the core column names of the
epidemiological parameter data exported by the \pkg{epireview} R package.
}
\keyword{datasets}
