% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param}
\alias{extract_param}
\title{Calculate the parameters of a parametric probability distribution from
reported values of percentiles, or median and range}
\usage{
extract_param(
  type = c("percentiles", "range"),
  values,
  distribution = c("lnorm", "gamma", "weibull"),
  percentiles,
  samples
)
}
\arguments{
\item{type}{A \code{character} defining whether summary statistics based
around \code{percentiles} (default) or \code{range}}

\item{values}{A \code{vector}. If \code{type = percentiles}: \code{c(percentile_1, percentile_2)}; and if \code{type = range}: \code{c(median, min, max)}}

\item{distribution}{A \code{character} specifying distribution to use.
Default is \code{lnorm}; also takes \code{gamma} and \code{weibull}.}

\item{percentiles}{A \code{vector} with two elements specifying the
percentiles defined in \code{values} if using \code{type = "percentiles"}.}

\item{samples}{A \code{numeric} specifying the sample size if using
\code{type = "range"}.}
}
\value{
A named \code{numeric} vector with the parameter values of the
distribution. If the \code{distribution = lnorm} then the parameters returned are
the meanlog and sdlog; if the \code{distribution = gamma} or \code{distribution = weibull} then the parameters returned are the shape and scale.
}
\description{
Summary data of distributions, as provided by reports and meta-analyses, can
be used to extract the parameters of a chosen distribution. Currently
available distributions are: lognormal, gamma and weibull. Extracting
from a lognormal returns the meanlog and sdlog parameters, and extracting
from the gamma and weibull returns the shape and scale parameters.
}
\examples{
# extract parameters of a lognormal distribution from the 75 percentiles
extract_param(
  type = "percentiles",
  values = c(6, 13),
  distribution = "lnorm",
  percentiles = c(0.125, 0.875)
)

# extract parameters of a gamma distribution from median and range
extract_param(
  type = "range",
  values = c(8, 6, 13),
  distribution = "gamma",
  samples = 20
)
}
\author{
Adam Kucharski
}
\keyword{extract}
