% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extract_param_percentile}
\alias{extract_param_percentile}
\title{Optimise the parameters for a specified probability distribution given the
percentiles of a distribution and the values at those percentiles}
\usage{
extract_param_percentile(values, distribution, percentiles)
}
\arguments{
\item{values}{A \code{vector}. If \code{type = percentiles}: \code{c(percentile_1, percentile_2)}; and if \code{type = range}: \code{c(median, min, max)}}

\item{distribution}{A \code{character} specifying distribution to use.
Default is \code{lnorm}; also takes \code{gamma} and \code{weibull}.}

\item{percentiles}{A \code{vector} with two elements specifying the
percentiles defined in \code{values} if using \code{type = "percentiles"}.
Percentiles should be specified between 0 and 1. For example 2.5th and 97.5th
percentile is given as \code{c(0.025, 0.975)}.}
}
\value{
A list with output from stats::optim. See ?optim for more details
}
\description{
Optimise the parameters for a specified probability distribution given the
percentiles of a distribution and the values at those percentiles
}
\keyword{internal}
