% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_param.R
\name{extraction_functions}
\alias{extraction_functions}
\alias{fit_function_lnorm_range}
\alias{fit_function_gamma_range}
\alias{fit_function_weibull_range}
\alias{fit_function_lnorm}
\alias{fit_function_gamma}
\alias{fit_function_weibull}
\title{Functions for extracting distribution parameters}
\usage{
fit_function_lnorm_range(param, val)

fit_function_gamma_range(param, val)

fit_function_weibull_range(param, val)

fit_function_lnorm(param, val)

fit_function_gamma(param, val)

fit_function_weibull(param, val)
}
\arguments{
\item{param}{Named numeric vector of the distribution parameters to be
optimised}

\item{val}{Numeric vector, in the case of using percentiles it contains the
values at the percentiles and the percentiles, in the case of median and
range it contains the median, lower range, upper range and the number of
sample points to evaluate the function at}
}
\description{
Set of functions that can be used to estimate the parameters of a
distribution (lognormal, gamma, Weibull) via optimisation from either the
percentiles or the median and ranges.
}
\author{
Adam Kucharski
}
\keyword{internal}
