% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{format.epiparam}
\alias{format.epiparam}
\title{Format method for epiparam class}
\usage{
\method{format}{epiparam}(x, ...)
}
\arguments{
\item{x}{epiparam object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Nothing (prints output)
}
\description{
Format method for epiparam class
}
\examples{
x <- epiparam()
format(x)
}
