% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{format.vb_epidist}
\alias{format.vb_epidist}
\title{Format method for \verb{<vb_epidist>} class}
\usage{
\method{format}{vb_epidist}(x, ...)
}
\arguments{
\item{x}{A \verb{<vb_epidist>} object.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
Invisibly returns a \verb{<vb_epidist>}. Called for printing
side-effects.
}
\description{
Format method for \verb{<vb_epidist>} class
}
\examples{
vb_epidist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "ebola",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1)
  ),
  extrinsic_epidist = epidist(
    disease = "ebola",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1)
  )
)
format(vb_epidist)
}
