% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation.multi_epidist}
\alias{get_citation.multi_epidist}
\title{Get citation from a list of \verb{<epidist>} objects}
\usage{
\method{get_citation}{multi_epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{Extra arguments to be passed to the method.}
}
\value{
A list of \verb{<bibentry>} objects. The length of output list is
equal to the length of the list of \verb{<epidist>} objects supplied.
}
\description{
Extract the citation stored in a list of \verb{<epidist>} objects.
}
\examples{
# example with list of <epidist>
edist <- epidist_db()
get_citation(edist)
}
