% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation}
\alias{get_citation}
\title{Extract citation information from \verb{<epidist>} or \verb{<epiparam>} objects}
\usage{
get_citation(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} or \verb{<epiparam>} object.}

\item{...}{Extra arguments to be passed to the method.}
}
\value{
A single character string or list of character string citations.
Length of list output is equal to number of rows in the \verb{<epiparam>} object
passed to the function.
}
\description{
Extract citation information from \verb{<epidist>} or \verb{<epiparam>} objects
}
\examples{

# example with epidist
eparam <- epiparam()
edist <- as_epidist(eparam[12, ])
get_citation(edist)

# example with epiparam
eparam <- epiparam()
get_citation(eparam)
}
