% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_citation}
\alias{get_citation}
\title{Get citation from an object}
\usage{
get_citation(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\description{
Extract the citation stored in an \R object.
}
