% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from an object}
\usage{
get_parameters(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\description{
Extract the parameters stored in an \R object.
}
