% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{head.epiparam}
\alias{head.epiparam}
\title{head method for epiparam class}
\usage{
\method{head}{epiparam}(x)
}
\arguments{
\item{x}{epiparam object}
}
\value{
Data frame
}
\description{
head method for epiparam class
}
\examples{
head(epiparam())
}
