% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{head.epiparam}
\alias{head.epiparam}
\alias{tail.epiparam}
\title{\code{head} and \code{tail} methods for \code{\link{epiparam}} class}
\usage{
\method{head}{epiparam}(x, ...)

\method{tail}{epiparam}(x, ...)
}
\arguments{
\item{x}{An \code{\link{epiparam}} object}

\item{...}{further arguments passed to or from other methods}
}
\value{
Data frame
}
\description{
\code{head} and \code{tail} methods for \code{\link{epiparam}} class
}
\examples{
head(epiparam())
tail(epiparam())
}
