% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{is_epidist}
\alias{is_epidist}
\title{Check object is an \verb{<epidist>}}
\usage{
is_epidist(x)
}
\arguments{
\item{x}{An \R object.}
}
\value{
A boolean logical, \code{TRUE} if the object is an \verb{<epidist>} and \code{FALSE}
if not.
}
\description{
Check object is an \verb{<epidist>}
}
\examples{
edist <- epidist(
  disease = "ebola",
  epi_dist = "serial_interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

is_epidist(edist)

false_edist <- list(
  disease = "ebola",
  epi_dist = "serial_interval",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)

is_epidist(false_edist)
}
