% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist_utils.R
\name{is_epidist_params}
\alias{is_epidist_params}
\title{Check if parameters are valid for an \code{epidist} object}
\usage{
is_epidist_params(prob_dist_params)
}
\arguments{
\item{prob_dist_params}{A named vector of probability distribution
parameters}
}
\value{
A logical boolean
}
\description{
A helper function to check whether the vector of parameters for the
probability distribution are in the set of possible parameters used in
the \code{epiparameter} package
}
\examples{
is_epidist_params(prob_dist_params = c(shape = 2, scale = 1))
}
