% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{is_epiparam}
\alias{is_epiparam}
\title{Check whether the object is an \code{epiparam}}
\usage{
is_epiparam(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A boolean logical, \code{TRUE} if the object is an \code{epiparam} and \code{FALSE}
if not
}
\description{
Check whether the object is an \code{epiparam}
}
\examples{
eparam <- epiparam()

is_epiparam(eparam)
}
