% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{is_parameterised}
\alias{is_parameterised}
\title{Checks whether \if{html}{\out{<epidist>}} object contains a distribution and parameters for
that distribution.}
\usage{
is_parameterised(x)
}
\arguments{
\item{x}{An \code{epidist} object.}
}
\value{
A boolean logical.
}
\description{
If the \if{html}{\out{<epidist>}} object is missing either a probability
distribution or parameters for the probability distribution,
\code{is_parameterised} returns \code{FALSE}, otherwise it returns \code{TRUE}
}
\examples{
# parameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
is_parameterised(edist)

# unparameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation"
)
is_parameterised(edist)
}
