% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_parameterised}
\alias{is_parameterised}
\alias{is_parameterized}
\title{Check if \verb{<epidist>} or list of \verb{<epidist>} objects contains a distribution
and distribution parameters}
\usage{
is_parameterised(x, ...)

is_parameterized(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} or list of \verb{<epidist>} objects.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
A single boolean \code{logical} for \verb{<epidist>} or vector of \code{logical}s
equal in length to the list of \verb{<epidist>} objects input. If the \verb{<epidist>}
object is missing either a probability distribution or parameters for
the probability distribution returns \code{FALSE}, otherwise it returns \code{TRUE}.
}
\description{
Check if \verb{<epidist>} or list of \verb{<epidist>} objects contains a distribution
and distribution parameters
}
\examples{
# parameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
is_parameterised(edist)

# unparameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation"
)
is_parameterised(edist)

# list of <epidist>
edist <- epidist_db()
is_parameterised(edist)
}
