% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{is_parameterised}
\alias{is_parameterised}
\title{Check if \verb{<epidist>} or \verb{<epiparam>} object contain a distribution and
distribution parameters}
\usage{
is_parameterised(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} or \verb{<epiparam>} object.}

\item{...}{\link{dots} not used, extra arguments supplied will cause a warning.}
}
\value{
A single boolean \code{logical} for \verb{<epidist>} or vector of boolean
\code{logical}s with length equal to the number of rows in the \verb{<epiparam>}.
If the \verb{<epidist>} object or a row in the \verb{<epiparam>} is missing either
a probability distribution or parameters for the probability distribution
returns \code{FALSE}, otherwise it returns \code{TRUE}.
}
\description{
Check if \verb{<epidist>} or \verb{<epiparam>} object contain a distribution and
distribution parameters
}
\examples{
# parameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
is_parameterised(edist)

# unparameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation"
)
is_parameterised(edist)
}
