% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List distributions}
\usage{
list_distributions(
  delay_dist = c("all", "incubation", "onset_to_admission", "onset_to_death",
    "serial_interval", "generation_time"),
  parameters = FALSE
)
}
\arguments{
\item{delay_dist}{A character defining parameter to be listed: \code{"all"}
\code{"incubation"}, \code{"onset_to_admission"}, \code{"onset_to_death"}, or
\code{"serial_interval"}. \code{"all"} will not filter by a certain delay distribution
and return all records. \code{"all"} is the default.}

\item{parameters}{A logical defining whether to show parameter values,
default is \code{FALSE}}
}
\description{
List distributions included in package
}
\examples{
\dontrun{
# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions()

# metrics can be chosen using the `delay_dist` argument, for example
# incubation period or onset to death
list_distributions(delay_dist = "incubation")
list_distributions(delay_dist = "onset_to_death")

# by default list_distributions() will return a filtered table, but if the
# full table including extra information (e.g. parameters of the
# distribution) the `parameters` argument can be set to `TRUE`
list_distributions(delay_dist = "onset_to_admission", parameters = TRUE)
}
}
\author{
Adam Kucharski
}
\keyword{distributions}
