% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List epidemiological distributions stored in the epiparameter library}
\usage{
list_distributions(disease = "all", epi_dist = "all", ..., multi_epidist)
}
\arguments{
\item{disease}{A \code{character} string with name of the infectious disease.}

\item{epi_dist}{A \code{character} string with the name of the
epidemiological distribution type.}

\item{...}{\link{dots} Extra arguments to be passed to \code{\link[=epidist_db]{epidist_db()}}.}

\item{multi_epidist}{Either an \verb{<epidist>} object or a list of \verb{<epidist>}
objects.}
}
\description{
This function subsets the epidemiological parameter library to
return only the chosen epidemiological distribution. The results are
returned as a data frame containing the disease, epidemiological
distribution, probability distribution, author of the study, and the year
of publication.
}
\details{
The \verb{<data.frame>} returned by \code{\link[=list_distributions]{list_distributions()}} contains some
vector columns and some list columns. For example, the \code{author} column is a
list as each paper can contain multiple authors.
}
\examples{
list_distributions(disease = "COVID-19", epi_dist = "incubation_period")

# example using list of `<epidist>` objects in memory
epidist_lst <- epidist_db(disease = "COVID-19")
list_distributions(multi_epidist = epidist_lst)

# example filtering an existing list to incubation periods
epidist_lst <- epidist_db(disease = "COVID-19")
list_distributions(
  multi_epidist = epidist_lst,
  epi_dist = "incubation period"
)
}
\author{
Adam Kucharski, Joshua W. Lambert
}
