% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_distributions.R
\name{list_distributions}
\alias{list_distributions}
\title{List distributions}
\usage{
list_distributions(delay_dist = "incubation", parameters = FALSE)
}
\arguments{
\item{delay_dist}{A \code{character} defining parameter to be listed: "incubation",
"onset_to_admission" or "onset_to_death"}

\item{parameters}{A \code{logical} defining whether to show parameter values}
}
\description{
List distributions included in package
}
\examples{
# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions()

# metrics can be chosen using the `delay_dist` argument, for example incubation
# period or onset to death
list_distributions(delay_dist = "incubation")
list_distributions(delay_dist = "onset_to_death")

# by default list_distributions() will return a filtered table, but if the
# full table including extra information (e.g. parameters of the
# distribution) the `parameters` argument can be set to `TRUE`
list_distributions(delay_dist = "onset_to_admission", parameters = TRUE)
}
\author{
Adam Kucharski
}
\keyword{distributions}
