% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{lnorm_meanlogsdlog2meansd}
\alias{lnorm_meanlogsdlog2meansd}
\title{Convert the meanlog and sdlog parameters of the lognormal distribution to
the mean and standard deviation}
\usage{
lnorm_meanlogsdlog2meansd(meanlog, sdlog)
}
\arguments{
\item{meanlog}{The meanlog parameter (mean of natural logarithm) of the
lognormal distribution. Also commonly called and denoted by the mu.}

\item{sdlog}{The sdlog parameter (standard deviation of the natural
logarithm) of the distribution. Also commonly called and denoted by sigma.}
}
\value{
A named list with mean and standard deviation
}
\description{
Convert the meanlog and sdlog parameters of the lognormal distribution to
the mean and standard deviation
}
\examples{
lnorm_meanlogsdlog2meansd(1.5, 0.9)
}
