% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{make_epidist}
\alias{make_epidist}
\title{Creates an \code{epidist} object from a list of input from an \code{epiparam} object}
\usage{
make_epidist(x)
}
\arguments{
\item{x}{List of data to be used to construct an epidist object}
}
\value{
An \code{epidist} object
}
\description{
Unpacks list of inputs from an epiparam object into the epidist
helper, including the parameters and uncertainty from the correct type of
probability distribution
}
\keyword{internal}
