% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{new_epidist}
\alias{new_epidist}
\title{Constructor for epidist class}
\usage{
new_epidist(
  disease = list(),
  epi_dist = character(),
  prob_dist = list(),
  prob_dist_params = numeric(),
  uncertainty = list(),
  summary_stats = list(),
  auto_calc_params = logical(),
  citation = character(),
  metadata = list(),
  method_assess = list(),
  discretise = logical(),
  truncation = numeric(),
  notes = character()
)
}
\arguments{
\item{disease}{A list contains the \verb{$disease} a character string of the
infectious disease specified in the study, and the \verb{$pathogen} a character
string. If the pathogen is unknown it can be given as \code{NULL}.}

\item{epi_dist}{A character string with the name of the
epidemiological distribution type}

\item{prob_dist}{A character string specifying the probability
distribution}

\item{prob_dist_params}{A named vector of probability distribution
parameters}

\item{uncertainty}{A list of named vectors with the uncertainty around
the probability distribution parameters. If uncertainty around the parameter
estimates is unknown use \code{create_epidist_uncertainty()} (which is the
argument default) to create a list wiht the correct names with missing
values.}

\item{summary_stats}{A list of summary statistics, use
\code{\link[=create_epidist_summary_stats]{create_epidist_summary_stats()}} to create list. This list can include
summary statistics about the inferred distribution such as it's mean and
standard deviation, quantiles of the distribution, or information about the
data used to fit the distribution such as lower and upper range. The summary
statistics can also include uncertainty around metrics such as confidence
interval around mean and standard deviation.}

\item{auto_calc_params}{A boolean logical determining whether to try and
calculate the probability distribution parameters from summary statistics if
distribution parameters are not provided. Default is \code{TRUE}. In the case when
sufficient summary statistics are provided and the parameter(s) of the
distribution are not, the \code{\link[=calc_dist_params]{calc_dist_params()}} function is called to
calculate the parameters and add them to the \code{epidist} object created.}

\item{citation}{A character string with the citation of the source of the
data or the paper that inferred the distribution parameters, use
\code{create_epidist_citation()} to create citation.}

\item{metadata}{A list of metadata, this can include: sample size, the
transmission mode of the disease (e.g. is it vector-borne or directly
transmitted), etc. It is assumed that the disease is not
vector-borne and that the distribution is intrinsic (e.g. not an extrinsic
delay distribution such as extrinsic incubation period) unless
\code{tranmission_mode = "vector_borne"} is contained in the metadata. Use
\code{create_epidist_metadata()} to create metadata.}

\item{method_assess}{A list of methodological aspects used when fitting
the distribution, use \code{create_epidist_method_assess()} to create method
assessment.}

\item{discretise}{A boolean logical whether the distribution is discretised.
Default is FALSE which assumes a continuous probability distribution}

\item{truncation}{A numeric specifying the truncation point if the inferred
distribution was truncated, NA if not or unknown.}

\item{notes}{A character string with any additional information about the
data, inference method or disease.}
}
\value{
epidist object
}
\description{
Creates an epidist object. The
constructor will search whether parameters of the probability distribution
are supplied and if not look to see whether they can be inferred/extracted/
converted from summary statistics provided. It will also convert the
probability distribution (\code{prob_dist}) and its parameters
(\code{prob_dist_params}) into an S3 class, either a \code{distribution} object from
\code{{distributional}} when \code{discretise = FALSE}, or a \code{distcrete} object from
\code{{distcrete}} when \code{discretise = TRUE}.
}
\examples{
epiparameter:::new_epidist(
  disease = list(disease = "ebola", pathogen = "ebola_virus"),
  epi_dist = "incubation_period",
  prob_dist = "gamma",
  prob_dist_params = c(shape = 1, scale = 1),
  uncertainty = create_epidist_uncertainty(),
  summary_stats = create_epidist_summary_stats(),
  auto_calc_params = TRUE,
  citation = create_epidist_citation(),
  metadata = create_epidist_metadata(),
  method_assess = create_epidist_method_assess(),
  discretise = FALSE,
  truncation = NA,
  notes = "No notes"
)
}
\keyword{internal}
