% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{new_epiparam}
\alias{new_epiparam}
\title{epiparam constructor}
\usage{
new_epiparam(epi_dist = character(), subset_params = logical())
}
\value{
epiparam object
}
\description{
The constructor reads the data stored internally in the package
and subsets by epidemiological distribution (epi_dist) or specifies what are
considered important information (subset_params)
}
\examples{
obj <- new_epiparam()
}
\keyword{internal}
