% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{new_epiparam}
\alias{new_epiparam}
\title{epiparam constructor}
\usage{
new_epiparam(epi_dist)
}
\value{
epiparam object
}
\description{
The constructor reads the data stored internally in the package
and subsets by epidemiological distribution (epi_dist)
}
\examples{
eparam <- epiparameter:::new_epiparam("all")
}
\keyword{internal}
