% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{new_vb_epidist}
\alias{new_vb_epidist}
\title{Constructor for \verb{<vb_epidist>} class}
\usage{
new_vb_epidist(intrinsic_epidist, extrinsic_epidist)
}
\arguments{
\item{intrinsic_epidist}{An \verb{<epidist>} object.}

\item{extrinsic_epidist}{An \verb{<epidist>} object.}
}
\value{
\verb{<vb_epidist>} object
}
\description{
Create an \verb{<vb_epidist>} object by binding two
\verb{<epidist>} objects and assigning the \verb{<vb_epidist>} class.
}
\keyword{internal}
