% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{parameters.epidist}
\alias{parameters.epidist}
\title{Parameters method for \code{epidist} object}
\usage{
\method{parameters}{epidist}(x, ...)
}
\arguments{
\item{x}{An \code{epidist} object.}

\item{...}{Additional arguments used by methods.}
}
\value{
A named vector of parameters or \code{NA} when the \verb{<epidist>} object is
unparameterised
}
\description{
Extracts the parameters of the distribution stored in an \code{epidist} object.
}
\details{
The \if{html}{\out{<epidist>}} object can be unparameterised in which it lacks
a probability distribution or parameters of a probability distribution.
In this can the \code{parameters.epidist()} method with return \code{NA}.
}
