% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epidist}
\alias{plot.epidist}
\title{Plot method for \verb{<epidist>} class}
\usage{
\method{plot}{epidist}(x, day_range = NULL, cumulative = FALSE, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{day_range}{Either \code{NULL} or a \code{numeric} vector of length 2 with the
first and last day to plot on the x-axis. If \code{NULL} the distribution is
plotted between day 0 and the 99th quantile of the distribution.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plot an \verb{<epidist>} object by displaying the either the
probability mass function (PMF), (in the case of discrete distributions)
or probability density function (PDF) (in the case of continuous
distributions), or the cumulative distribution function (CDF).
}
\examples{
# plot continuous epidist
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 2, scale = 1)
)
plot(edist)

# plot different day range (x-axis)
plot(edist, day_range = c(0, 10))

# plot CDF
plot(edist, cumulative = TRUE)

# plot discrete epidist
edist <- discretise(edist)
plot(edist)
}
\author{
Joshua W. Lambert
}
