% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{plot.epidist}
\alias{plot.epidist}
\title{PLots an \code{epidist} object}
\usage{
\method{plot}{epidist}(x, day_range = 0:10, ..., vb = FALSE, title = NULL)
}
\arguments{
\item{x}{An \code{epidist} object}

\item{day_range}{A vector with the sequence of days to be plotted on the
x-axis of the distribution}

\item{...}{Allow other graphical parameters}

\item{vb}{A boolean logical determining whether the \code{epidist} being plotted
has come from a \code{vb_epidist} object}

\item{title}{Either a character string or \code{NULL}. If not null the character
string will be printed as a title to the plot}
}
\description{
Plots an \code{epidist} object by displaying the either the
probability mass function (PMF), (in the case of discrete distributions)
or probability density function (PDF) (in the case of continuous
distributions) and the cumulative distribution function (CDF). Resulting in
a 1x2 grid plot.
}
\examples{
# plot continuous epidist
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 2, scale = 1)
)

plot(edist, day_range = 0:10)
# plot discrete epidist
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 2, scale = 1),
  discretise = TRUE
)

plot(edist, day_range = 0:10)
}
\author{
Joshua W. Lambert
}
