% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{plot.epidist}
\alias{plot.epidist}
\title{Plots an \code{epidist} object by displaying the probability mass function (PMF),
probability density function (PDF) and cumulative distribution function (CDF)}
\usage{
\method{plot}{epidist}(x, day_range = 0:10, ...)
}
\arguments{
\item{x}{An \code{epidist} object}

\item{day_range}{A vector with the sequence of days to be plotted on the
x-axis of the distribution}

\item{...}{Allow other graphical parameters}
}
\description{
Plots an \code{epidist} object by displaying the probability mass function (PMF),
probability density function (PDF) and cumulative distribution function (CDF)
}
\examples{
\dontrun{
ebola_si <- epidist(pathogen = "ebola", delay_dist = "serial_interval")
plot(ebola_si)
}
}
\author{
Joshua W. Lambert
}
