% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{plot.vb_epidist}
\alias{plot.vb_epidist}
\title{Plots an \code{vb_epidist} object,}
\usage{
\method{plot}{vb_epidist}(x, day_range = 0:10, ...)
}
\arguments{
\item{x}{An \code{epidist} object}

\item{day_range}{A vector with the sequence of days to be plotted on the
x-axis of the distribution}

\item{...}{Allow other graphical parameters}
}
\description{
Plots a \code{vb_epidist} object by displaying the either the
probability mass function (PMF), (in the case of discrete distributions)
or probability density function (PDF) (in the case of continuous
distributions) and the cumulative distribution function (CDF), for both the
intrinsic and extrinsic distributions. This resulting in a 2x2 grid plot.
}
\examples{
# plot vb_epidist
dengue_dist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      vector_borne = TRUE,
      extrinsic = FALSE
    )
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      vector_borne = TRUE,
      extrinsic = TRUE
    )
  )
)

plot(dengue_dist, day_range = 0:10)
}
\author{
Joshua W. Lambert
}
