% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{print.epidist}
\alias{print.epidist}
\title{Print method for epidist class}
\usage{
\method{print}{epidist}(x, header = TRUE, vb = NULL)
}
\arguments{
\item{x}{epidist object}

\item{header}{Boolean logical determining whether the header (first part) of
the print method is printed. This is used internally for plotting the
vb_epidist class}

\item{vb}{A character string containing whether it is the intrinsic
(\code{"Intrinsic"}) or extrinsic (\code{"extrinsic"}) distribution for vector-borne
diseases}
}
\value{
Nothing (prints output)
}
\description{
Print method for epidist class
}
\examples{
epidist <- epidist(
  disease = "ebola",
  epi_distribution = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
epidist
}
