% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{print.epidist}
\alias{print.epidist}
\title{Print method for \verb{<epidist>} class}
\usage{
\method{print}{epidist}(x, ...)
}
\arguments{
\item{x}{An \verb{<epidist>} object.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}
}
\value{
Invisibly returns an \verb{<epidist>}. Called for side-effects.
}
\description{
Print method for \verb{<epidist>} class
}
\examples{
epidist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
epidist
}
