% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{print.epiparam}
\alias{print.epiparam}
\title{Print method for \verb{<epiparam>} class}
\usage{
\method{print}{epiparam}(x, ...)
}
\arguments{
\item{x}{An \verb{<epiparam>} object.}

\item{...}{\link{dots} Extra arguments passed to or from other methods.}
}
\value{
Invisibly returns an \verb{<epiparam>}. Called for side-effects.
}
\description{
Print method for \verb{<epiparam>} class
}
\examples{
epiparam <- epiparam()
epiparam
}
