% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.multi_epidist}
\alias{print.multi_epidist}
\title{Print method for \verb{<multi_epidist>} class}
\usage{
\method{print}{multi_epidist}(x, ..., n = NULL)
}
\arguments{
\item{x}{A \verb{<multi_epidist>} object.}

\item{...}{\link{dots} Extra arguments to be passed to the method.}

\item{n}{A \code{numeric} specifying how many \verb{<epidist>} objects to print. This
argument is passed to \code{\link[=head]{head()}} for \code{list} printing. Default is \code{NULL} and the
number of elements to print is controlled by package \code{\link[=options]{options()}}.}
}
\value{
Invisibly returns a \verb{<multi_epidist>}. Called for side-effects.
}
\description{
Print method for \verb{<multi_epidist>} class
}
