% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{print.vb_epidist}
\alias{print.vb_epidist}
\title{Print method for vb_epidist class}
\usage{
\method{print}{vb_epidist}(x, ...)
}
\arguments{
\item{x}{vb_epidist object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Invisibly returns a \code{\link{vb_epidist}}. Called for printing
side-effects.
}
\description{
Print method for vb_epidist class
}
\examples{
vb_epidist <- vb_epidist(
  intrinsic = epidist(
    disease = "ebola",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1)
  ),
  extrinsic = epidist(
    disease = "ebola",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1)
  )
)
vb_epidist
}
