% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam_utils.R
\name{[.epiparam}
\alias{[.epiparam}
\title{Subsetting method for \code{epiparam} class that fails when epiparam object is
invalidated}
\usage{
\method{[}{epiparam}(epiparam, ...)
}
\arguments{
\item{epiparam}{An \code{epiparam} object}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
An \code{epiparam} object
}
\description{
Subsetting method for \code{epiparam} class that fails when epiparam object is
invalidated
}
