% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{summary.epiparam}
\alias{summary.epiparam}
\title{Summary method for epiparam class}
\usage{
\method{summary}{epiparam}(x)
}
\arguments{
\item{x}{epiparam object}
}
\value{
data frame of information
}
\description{
Summary method for epiparam class
}
\examples{
x <- epiparam()
summary(x)
}
