% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidist.R
\name{validate_epidist}
\alias{validate_epidist}
\title{Validator for \verb{<epidist>} class}
\usage{
validate_epidist(epidist)
}
\arguments{
\item{epidist}{An \verb{<epidist>} object}
}
\value{
Invisibly returns an \verb{<epidist>}. Called for side-effects (errors
when invalid \verb{<epidist>} object is provided).
}
\description{
Validator for \verb{<epidist>} class
}
