% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiparam.R
\name{validate_epiparam}
\alias{validate_epiparam}
\title{\code{epiparam} class validator}
\usage{
validate_epiparam(epiparam)
}
\arguments{
\item{epiparam}{An \code{epiparam} object}
}
\value{
Invisibly returns an \code{\link{epiparam}}. Called for side-effects
(errors when invalid \code{epiparam} object is provided).
}
\description{
\code{epiparam} class validator
}
