% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vb_epidist.R
\name{validate_vb_epidist}
\alias{validate_vb_epidist}
\title{\code{vb_epidist} class validator}
\usage{
validate_vb_epidist(vb_epidist)
}
\arguments{
\item{vb_epidist}{A \code{vb_epidist} object.}
}
\value{
Invisibly returns a \code{\link{vb_epidist}}. Called for side-effects (errors
when invalid \code{vb_epidist} object is provided).
}
\description{
\code{vb_epidist} class validator
}
