% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{weibull_shapescale2meansd}
\alias{weibull_shapescale2meansd}
\title{Converts the shape and scale parameters of the Weibull distribution to the
mean and standard deviation}
\usage{
weibull_shapescale2meansd(shape, scale)
}
\arguments{
\item{shape}{The shape parameter of the Weibull distribution}

\item{scale}{The scale parameter of the Weibull distribution}
}
\value{
A named list with mean and standard deviation
}
\description{
Converts the shape and scale parameters of the Weibull distribution to the
mean and standard deviation
}
\examples{
weibull_shapescale2meansd(shape = 2, scale = 1)
}
